### 2024-11-12 #7467 @editorialbot
[PRE REVIEW]: textNet: Directed, Multiplex, Multimodal Event Network Extraction from Textual Data
**Submitting author:** @ezufall (<a href="http://orcid.org/0000-0003-4179-1163">Elise Zufall</a>)
**Repository:** https://github.com/ucd-cepb/textNet
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
