### 2024-12-05 4ead490
acceptance of null custom entities
