### 2024-12-24 #7615 @mikemahoney218
👋🏼 @ezufall, @amatsuo, @stuartyeates: this is the review thread for the paper. Just about all of our communications will happen here from now on :smile:

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4 weeks, though obviously with the holidays we're expecting a slightly longer timeframe. Please let me know if you require some more time!
