### 2024-12-30 #7615 @mikemahoney218
Hi all -- @kbenoit has graciously agreed to be our third reviewer! While we can accept a submission on the basis of two reviews, I always prefer to have a third. So, thanks again @kbenoit , and there's a few more instructions over at
https://github.com/openjournals/joss-reviews/issues/7615#issuecomment-2561124751
