### 2025-01-06 a6810fc
accommodate empty appositive list and empty edgelist, and AUX pos
