+++ R/textnet_extract.R
@@ -110,7 +110,7 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
   dep_rels_subj_keep <- c('nsubj','nsubjpass','csubj','csubjpass','agent','expl')        
   dep_rels_obj_keep <- c('pobj','iobj','dative','attr','dobj','oprd','ccomp','xcomp','acomp','pcom') 
 
-  x <- x[,keep:=any(dep_rel %in% dep_rels_subj_keep) & any(dep_rel %in% dep_rels_obj_keep) & any(pos=='VERB'),by=.(doc_id,sentence_id)]
+  x <- x[,keep:=any(dep_rel %in% dep_rels_subj_keep) & any(dep_rel %in% dep_rels_obj_keep) & any(pos%in%c('VERB','AUX')),by=.(doc_id,sentence_id)]
   x <- x[keep==T,]
   x$doc_sent <- paste0(x$doc_id, "_", x$sentence_id)
 
@@ -150,8 +150,16 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
       }
   }, cl=cl)
   apposlist <- data.table::rbindlist(apposlist)
-  apposlist <- base::unique(apposlist[nchar(apposlist$fullname)>0,])
-  colnames(apposlist) <- c("abbrev", "fullname")
+  
+  #if it's not empty, it will have two columns, which we can name.
+  if(ncol(apposlist)>0){
+    colnames(apposlist) <- c("abbrev", "fullname")
+    apposlist <- base::unique(apposlist[nchar(apposlist$fullname)>0,])
+  #if it's empty, we create an empty 2-column dataframe
+  }else{
+    apposlist <- data.table::data.table("abbrev" = character(), "fullname" = character())
+  }
+
     
   #we've already incorporated aux and xcomp helpers into the verb phrase as edge attributes
   #so they should not count as their own row
@@ -235,9 +243,28 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
   #create all combos of source and target by doc_sent_verb
   st_pivot <- data.table::dcast(source_target_list,doc_sent_verb+row_id~source_or_target, value.var= "entity_name")
   
-  st_pivot <- data.table::as.data.table(st_pivot %>% dplyr::group_by(doc_sent_verb) %>% tidyr::expand(source, target))
-
-  edgelist <- data.table::merge.data.table(st_pivot, verb_dt, by = c("doc_sent_verb"),all.x=F, all.y=F)
+  if(nrow(st_pivot)>0){
+    st_pivot <- data.table::as.data.table(st_pivot %>% dplyr::group_by(doc_sent_verb) %>% tidyr::expand(source, target))
+    edgelist <- data.table::merge.data.table(st_pivot, verb_dt, by = c("doc_sent_verb"),all.x=F, all.y=F)
+    
+  }else{
+    #make an empty data.table if there are no edges in the network
+    edgelist <- data.table::data.table("source" = character(),
+                                       "target" = character(),
+                                       "head_verb_name" = character(),
+                                       "head_verb_lemma" = character(),
+                                       "head_verb_tense" = character(),
+                                       "helper_lemma" = list(),
+                                       "helper_token" = list(),
+                                       "xcomp_verb" = list(),
+                                       "xcomp_helper_lemma" = list(),
+                                       "xcomp_helper_token" = list(),
+                                       "neg" = logical(),
+                                       "edgeiscomplete" = logical(),
+                                       "has_hedge" = logical(),
+                                       "is_future" = logical(),
[…]
