### 2025-01-09 #7615 @stuartyeates
Looking over the repo (haven't downloaded or run the code yet):

(a) Two of the three PDFs appear to have been generated using a technique which means they can't be searched. While not a fatal flaw, this makes the contained information considerably less useful and the project as a whole less transparent.

(b) There appears to be no explicit mention that that all code appears to only have been tested on English-language content. Working in an English-only context is not a fatal flaw, but needs to be mentioned as a limitation.

(c) The repo's one line "About" field is empty. This is used by third party tools which will be marginally less useful without this. Copy the first content line from the README.

More feedback to come.
