### 2025-01-14 #7615 @stuartyeates
Code isn't installing cleanly for me.  On a fresh install of R as packaged by my institution:

R version 4.3.1 (2023-06-16 ucrt) -- "Beagle Scouts"
Copyright (C) 2023 The R Foundation for Statistical Computing
Platform: x86_64-w64-mingw32/x64 (64-bit)

I needed to manually:

install.packages("reticulate")
install.packages("spacyr")

before the

devtools::install_github('ucd-cepb/textNet')

------------------------------------------------

The PDFs

https://github.com/ucd-cepb/textnet/blob/main/vignettes/test_procedures.pdf
https://github.com/ucd-cepb/textnet/blob/main/vignettes/textNet_vignette_2024.pdf

have code examples which are both too long to reliably re-key and not amenable to cutting and pasting from the PDF (at least on my platform Windows / firefox), making them effectively useless.
