### 2025-01-14 e746676
manually ad pdfs
for some reason, pdfs are rendering in temp folder, not correct folder
