+++ vignettes/textNet_vignette_2024.Rmd
@@ -8,6 +8,7 @@ output: pdf_document
 
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
+knitr::opts_knit$set(root.dir = "~/Documents")
 ```
 
 ## Introduction
