### 2025-03-25 860def4
edits to documentation in response to reviewer feedback
