### 2025-03-25 #7615 @ezufall
Hello all, I likewise was unclear on the sequencing and thought perhaps all reviewers would comment in a single round.

I'm happy to provide an update and apologize for the delay.

We thank Dr. Yeates for the constructive feedback. In response to point (a), we re-rendered the vignette pdfs, which are now amenable to copy and paste. We also thank Dr. Yeates (b) for the recommendation to clarify the scope of our testing, which is currently in English, (c) the recommendation to include README content in the "About" field of the repo, and (d) the need to mention the primary platform (R) in the README. We have now amended the README and "About" field to incorporate points (b), (c), and (d).

We are also grateful for the opportunity to improve the installation instructions. The textNet package suggests, rather than requires, the spacyr package, as it provides an optional wrapper to assist with preprocessing before use of the main tool textnet_extract(). We have added clarification of the installation procedures to the README.
