### 2025-03-25 e1a49eb
custom entity adjustment to inside main loop
