### 2025-04-09 ucd-cepb/textnet#13 @amatsuo
Code chunks improvement in vignette and test procedures
## Code comments are too sparce in vignette and test procedures

- The code chunks in the vignette contain very few comments, which makes it harder for new readers to follow what each line is doing.
- Adding brief but meaningful comments to each code chunk (especially in the usage and example sections) would improve readability.

## Inconsistent indentation in vignette

- It seems that you want to make all code indented by 4-spaces, is it necessary?
- It may be fine to do so if it's appliced consistently. But it's not. This can make the code look untidy and harder to read.

https://github.com/openjournals/joss-reviews/issues/7615
