### 2025-04-09 ucd-cepb/textnet#14 @amatsuo
Inconsistent results
The resulting figures in "Explore Node Attributes" subsection are different between vignette and test procedures. Why?

https://github.com/openjournals/joss-reviews/issues/7615
