### 2025-04-09 ucd-cepb/textnet#16 @amatsuo
Introduction/Summary improvement
The introduction could benefit from clearer motivation. Network measurement is relevant for some researchers but to capture the attention of the wider audience, providing concrete examples of research questions that are uniquely addressable with textNet's would strengthen this section. Maybe a brief summary of your papers would do the job.

https://github.com/openjournals/joss-reviews/issues/7615
