### 2025-07-17 ea8cbca
limit scope of recompile draft-pdf.yml
