+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,10 @@
 name: Draft PDF
-on: [push]
+on: 
+  push:
+    paths:
+      - vignettes/paper.md
+      - vignettes/paper.bib
+      - .github/workflows/draft-pdf.yml
 
 jobs:
   paper:
