### 2025-07-17 34705c9
resolved suggested package handling and divergent test output
