+++ NAMESPACE
@@ -51,13 +51,9 @@ importFrom(ohenery,normalize)
 importFrom(pbapply,pblapply)
 importFrom(pdftools,pdf_ocr_text)
 importFrom(pdftools,pdf_text)
-importFrom(reticulate,py_config)
 importFrom(sna,centralization)
 importFrom(sna,connectedness)
 importFrom(sna,gtrans)
-importFrom(spacyr,spacy_finalize)
-importFrom(spacyr,spacy_initialize)
-importFrom(spacyr,spacy_parse)
 importFrom(stats,aggregate)
 importFrom(stats,median)
 importFrom(stringi,stri_escape_unicode)

[…]
