### 2025-07-17 e03bfa0
added supporting research examples to introduction
