### 2025-07-17 c994b44
additional commenting to support new users
