+++ vignettes/textNet_vignette_2025.Rmd
@@ -75,16 +75,26 @@ First, we read in the pre-processed data and call textnet_extract() to produce t
 
 ```{r readpreprocessed, message=F, warning=F}
 library(textNet)
+#load sample tabular data, in which 
+#each row is a single token
 old_new_parsed <- textNet::old_new_parsed
 
 ```
 
 ```{r extract, results = 'hide'}
+#initialize empty list of extracts to populate
 extracts <- vector(mode="list",length=length(old_new_parsed))
-  for(m in 1:length(old_new_parsed)){
-    extracts[[m]] <- textnet_extract(old_new_parsed[[m]],concatenator="_",cl=4,
+#populate each element of list with results of textnet_extract
+#increasing "cl" to 4 for speed
+#keeping four types of entities: organizations, 
+#geopolitical entities, people, and water bodies
+#keep_incomplete_edges is F by default; T keeps
+#'edges' that are missing either a source node
+#or a target node
+for(m in 1:length(old_new_parsed)){
+  extracts[[m]] <- textnet_extract(old_new_parsed[[m]],cl=4,
                    keep_entities = c('ORG','GPE','PERSON','WATER'), 
-                   return_to_memory=T, keep_incomplete_edges=T)
+                   keep_incomplete_edges=T)
    }
    
 ```
@@ -100,8 +110,10 @@ The resulting object consists of a nodelist, an edgelist, a verblist, and a list
 In a document, the same real-world entity may be referenced in multiple ways. For instance, the document may introduce an organization using its full name, then use an acronym for the remainder of the document. To have more reliable network results, it is important to consolidate nodes that represent different naming conventions into a single node. The *textNet* package comes with a built-in tool for finding acronyms defined parenthetically within the text. This can be run on the result of pdf_clean() to generate a table with one column for acronyms and another for the corresponding full names, such that each row is a different instance of a phrase for which an acronym was detected. The use of find_acronyms() is demonstrated below.
 
 ```{r acronyms}
-
+#reading in sample text data
 old_new_text <- textNet::old_new_text
+#use find_acronyms feature to find and tabulate
+#parenthetically defined acronyms within the text
 old_acronyms <- find_acronyms(old_new_text[[1]])
 new_acronyms <- find_acronyms(old_new_text[[2]])
    
@@ -128,6 +140,8 @@ The textnet_extract argument of disambiguate() accepts the result of the textnet
 Information about the optional argument try_drop can be found in the package documentation. When specified, the function merges nodes that differ only by the regex phrase specified by try_drop, and which become identical upon removal of the regular expression encoded in try_drop.
 
 ```{r consolidate}
+#the user can define a many-to-many matching
+#scheme using a tofrom object as shown below
 tofrom <- data.table::data.table(
   from = c(as.list(old_acronyms$acronym),
              list("Sub_basin",
@@ -144,12 +158,16 @@ tofrom <- data.table::data.table(
                   c("upper_aquifer","lower_aquifer"),
                   c("upper_aquifer","lower_aquifer"))))
 
+#conduct many-to-many mapping and consolidate nodes that represent
+#multiple ways of referring to a single entity 
+#(e.g. acronyms. vs. full name)
 old_extract_clean <- disambiguate(
       textnet_extract = extracts[[1]],
       from = tofrom$from, 
       to = tofrom$to,
       match_partial_entity = c(rep(F,nrow(old_acronyms)),T,T,F,F,F,F))
-   
+
+#same thing, now for the new network
 tofrom <- data.table::data.table(
    from = c(as.list(new_acronyms$acronym),
              list("Sub_basin",
@@ -179,6 +197,8 @@ new_extract_clean <- disambiguate(
[…]
