+++ vignettes/test_procedures.Rmd
@@ -13,36 +13,31 @@ knitr::opts_chunk$set(echo = TRUE)
 ### Pre-Processing Step I: Process PDFs
 
 ```{r pdf_clean, message=F, warning=F}
-   library(textNet)
-   library(stringr)
-   library(testthat)
-   URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/2840"
-   download.file(URL, destfile = "old.pdf", method="curl")
-   
-   URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/9625"
-   download.file(URL, destfile = "new.pdf", method="curl")
-   
-   pdfs <- c("old.pdf", 
+library(textNet)
+library(stringr)
+library(testthat)
+#these example pdfs are included with the package
+pdfs <- c("old.pdf", 
           "new.pdf")
    
-   old_new_text <- textNet::pdf_clean(pdfs, ocr=F, maxchar=10000, 
+old_new_text <- textNet::pdf_clean(pdfs, ocr=F, maxchar=10000, 
                      export_paths=NULL, return_to_memory=T, suppressWarn = F, 
                      auto_headfoot_remove = T)
-   names(old_new_text) <- c("old","new")
+names(old_new_text) <- c("old","new")
    
-   #we expect one element per pdf
-   expect_that(length(old_new_text), equals(length(pdfs)))
+#we expect one element per pdf
+expect_that(length(old_new_text), equals(length(pdfs)))
 ```
 
 ### Pre-Processing Step II: Parse Text
 
 
 ```{r parse, message=F, warning=F}
-   library(findpython)
-   ret_path <- find_python_cmd(required_modules = c('spacy', 'en_core_web_lg'))
+library(findpython)
+ret_path <- find_python_cmd(required_modules = c('spacy', 'en_core_web_lg'))
    
 
-   water_bodies <- c("surface water", "Surface water", "groundwater", "Groundwater", 
+water_bodies <- c("surface water", "Surface water", "groundwater", "Groundwater", 
                   "San Joaquin River", "Cottonwood Creek", "Chowchilla Canal Bypass", 
                   "Friant Dam", "Sack Dam", "Friant Canal", "Chowchilla Bypass", 
                   "Fresno River", "Sacramento River", "Merced River","Chowchilla River", 
@@ -53,11 +48,11 @@ knitr::opts_chunk$set(echo = TRUE)
                   "Lower aquifer", "Upper aquifer", "Upper and lower aquifers", 
                   "Lower and upper aquifers")
 
-   if(!requireNamespace("spacyr", quietly = T)){
+if(!requireNamespace("spacyr", quietly = T)){
     stop("Package 'spacyr' must be installed to use this function.",
          call.=F)
   }
-   old_new_parsed <- textNet::parse_text(ret_path, 
+old_new_parsed <- textNet::parse_text(ret_path, 
                           keep_hyph_together = F, 
                           phrases_to_concatenate = water_bodies, 
                           concatenator = "_", 
@@ -66,9 +61,9 @@ knitr::opts_chunk$set(echo = TRUE)
                                  overwrite = T,
                           custom_entities = list(WATER = water_bodies))
    
-   #expect all pages are preserved 
-   for(k in 1:length(old_new_parsed)){
-     maxpage <- max(as.numeric(stringr::str_remove(old_new_parsed[[k]]$doc_id, "text")))
+#expect all pages are preserved 
+for(k in 1:length(old_new_parsed)){
+   maxpage <- max(as.numeric(stringr::str_remove(old_new_parsed[[k]]$doc_id, "text")))
    expect_that(maxpage, equals(length(old_new_text[[k]])))
    }
    
[…]
