### 2025-07-18 3ba7ebb
'suggests' handling and rproj update
