### 2025-07-18 1cbfb78
intro wording adjustments
