+++ vignettes/paper.md
@@ -33,7 +33,7 @@ multimodal network graphs, enabling rich analysis of the relationships between
 verb attributes and tenses, entity types, structural motifs, and other network 
 characteristics. For instance, [Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556)
 demonstrate the use of *textNet* to identify which actors are involved in start-up
-versus ongoing or future management tasks, characterize patterns of information 
+versus ongoing management tasks, characterize patterns of information 
 and funding flows, and compare the distribution of management tasks in networks from
 regions known to have contrasting characteristics [-@zufall_syntactic_2024]. 
 *textNet* also works on arbitrarily long documents, making it well suited 
@@ -42,7 +42,7 @@ court proceedings, regulatory impact analyses, and environmental impact assessme
 
 *textNet* has applications in governance network scholarship, as demonstrated by
 [Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556) 
-[-@zufall_syntactic_2024] and by ongoing work on groundwater governance at the 
+[-@zufall_syntactic_2024] and by ongoing work on water resources governance at the 
 UC Davis Center for Environmental Policy and Behavior. Additional potential 
 applications include legal scholarship, social-ecological network analysis, 
 government planning documents, court proceedings, archival research, communication 
