### 2025-07-18 ucd-cepb/textnet#16 @ezufall
thank you for the feedback -- we have now added additional support to the introduction as suggested
