### 2025-07-18 906d3a2
affiliation adjustment
