### 2025-07-18 2d15719
Update README.md with new vignette name
