### 2025-09-23 75d0e01
add trf spacy model and start entity ruler
