+++ R/entity_specify.R
@@ -1,0 +1,195 @@
+# Exported function
+#entity_specify
+
+#' Creates EntityRuler patterns from entity names
+#' @param entity_names A character vector of entity names to be recognized. If file_path is provided, this parameter is ignored.
+#' @param entity_label A character string specifying the entity label to assign to all entities. Defaults to "CUSTOM". Ignored when reading from file.
+#' @param case_sensitive A logical value indicating whether pattern matching should be case sensitive. Defaults to FALSE. Ignored when reading from file.
+#' @param whole_word_only A logical value indicating whether to match only whole words. Defaults to TRUE. Ignored when reading from file.
+#' @param aliases A character vector of aliases for each entity. Can be semicolon-separated strings for multiple aliases per entity. Defaults to NULL. Ignored when reading from file.
+#' @param file_path Optional. Path to a CSV or TXT file containing entity specifications. File must have columns: entity_names, entity_label, case_sensitive, whole_word_only, aliases.
+#' @return A list of pattern dictionaries formatted for spaCy's EntityRuler
+#' @importFrom utils read.csv read.table
+#' @export
+
+entity_specify <- function(entity_names = NULL, entity_label = "CUSTOM", case_sensitive = FALSE, whole_word_only = TRUE, aliases = NULL, file_path = NULL) {
+  
+  # Handle file input or direct parameter input
+  if(!is.null(file_path)) {
+    # File input mode
+    if(!is.character(file_path) || length(file_path) != 1) {
+      stop("'file_path' must be a single character string")
+    }
+    
+    if(!file.exists(file_path)) {
+      stop(paste0("File not found: ", file_path))
+    }
+    
+    # Determine file type and read accordingly
+    file_ext <- tolower(tools::file_ext(file_path))
+    
+    if(file_ext == "csv") {
+      entity_data <- utils::read.csv(file_path, stringsAsFactors = FALSE)
+    } else if(file_ext %in% c("txt", "tsv")) {
+      entity_data <- utils::read.table(file_path, header = TRUE, sep = "\t", stringsAsFactors = FALSE)
+    } else {
+      stop("Supported file formats: .csv, .txt, .tsv")
+    }
+    
+    # Validate required columns
+    required_cols <- c("entity_names", "entity_label", "case_sensitive", "whole_word_only", "aliases")
+    missing_cols <- setdiff(required_cols, names(entity_data))
+    if(length(missing_cols) > 0) {
+      stop(paste0("Missing required columns: ", paste(missing_cols, collapse = ", ")))
+    }
+    
+    # Validate data types
+    if(!is.character(entity_data$entity_names)) {
+      entity_data$entity_names <- as.character(entity_data$entity_names)
+    }
+    if(!is.character(entity_data$entity_label)) {
+      entity_data$entity_label <- as.character(entity_data$entity_label)
+    }
+    if(!is.logical(entity_data$case_sensitive)) {
+      entity_data$case_sensitive <- as.logical(entity_data$case_sensitive)
+    }
+    if(!is.logical(entity_data$whole_word_only)) {
+      entity_data$whole_word_only <- as.logical(entity_data$whole_word_only)
+    }
+    if(!is.character(entity_data$aliases)) {
+      entity_data$aliases <- as.character(entity_data$aliases)
+    }
+    # Replace NA aliases with empty strings
+    entity_data$aliases[is.na(entity_data$aliases)] <- ""
+    
+    # Remove rows with missing entity names
+    entity_data <- entity_data[!is.na(entity_data$entity_names) & nchar(entity_data$entity_names) > 0, ]
+    
+    if(nrow(entity_data) == 0) {
[…]
