### 2025-09-23 6efebc0
entity ruler check
