### 2025-11-04 #7615 @kbenoit
I surprised that this was submitted to _JOSS_ without even passing the standard CRAN checks. These need to be fixed before this can be considered for _JOSS_. Let's start with basic items on the checklist:

### LICENSE:

RCHECK produces:

Non-standard files/directories found at top level:
  ‘LICENSE.txt’ ‘Rbuildignore.txt’ ‘create_heximage.R’ ‘img’

You should delete `LICENSE.md` and `LICENSE.txt` since the `DESCRIPTION` file references only `LICENSE`.

To stop the other warnings, add these entries to `.Rbuildignore`:

```
^create_heximage\.R$
^img/
^Rbuildignore\.txt$
```

### Other CHECK problems

```
── R CMD check results ───────────────────────────────────────────────────────────────────────────────────────────────── textNet 1.0.1 ────
Duration: 18.8s

❯ checking Rd \usage sections ... WARNING
  Undocumented arguments in Rd file 'parse_text.Rd'
    ‘model’

  Functions with \usage entries need to have the appropriate \alias
  entries, and all their arguments documented.
  The \usage entries must correspond to syntactically valid R code.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.

❯ checking files in ‘vignettes’ ... WARNING
  Files in the 'vignettes' directory but no files in 'inst/doc':
    ‘new.pdf’ ‘old.pdf’ ‘paper.bib’ ‘paper.md’ ‘paper.pdf’
    ‘test_procedures.Rmd’ ‘test_procedures.pdf’
    ‘textNet_vignette_2025.Rmd’ ‘textNet_vignette_2025.pdf’
    ‘vignette_figures/figure-markdown_strict/composite-1.png’
    ‘vignette_figures/figure-markdown_strict/flowchart.png’
    ‘vignette_figures/figure-markdown_strict/plot-1.png’
    ‘vignette_figures/figure-markdown_strict/plot-2.png’
    ‘vignette_figures/figure-markdown_strict/step8b4-1.png’
  Files named as vignettes but with no recognized vignette engine:
     ‘vignettes/test_procedures.Rmd’
     ‘vignettes/textNet_vignette_2025.Rmd’
  (Is a VignetteBuilder field missing?)

❯ checking for hidden files and directories ... NOTE
  Found the following hidden files and directories:
    .github
  These were most likely included in error. See section ‘Package
  structure’ in the ‘Writing R Extensions’ manual.

❯ checking DESCRIPTION meta-information ... NOTE
  Licence stub is invalid DCF.

❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘LICENSE.txt’ ‘Rbuildignore.txt’ ‘create_heximage.R’ ‘img’

❯ checking R code for possible problems ... NOTE
  clean_entities : <anonymous>: warning in stri_replace_all_regex(v[i],
    pattern = maths, replacement = letts, vectorize = F): partial
    argument match of 'vectorize' to 'vectorize_all'
  combine_networks: no visible binding for global variable
    ‘num_appearances’
  combine_networks: no visible binding for global variable
    ‘num_graphs_in’
  crawl_sentence: no visible binding for global variable ‘head_token_id’
  crawl_sentence: no visible binding for global variable ‘token_id’
  crawl_sentence: no visible binding for global variable ‘token’
  crawl_sentence: no visible binding for global variable ‘lemma’
  crawl_sentence: no visible binding for global variable ‘tag’
  crawl_sentence: no visible binding for global variable ‘dep_rel’
  crawl_sentence: no visible binding for global variable ‘pos’
  disambiguate: no visible binding for global variable ‘num_appearances’
  disambiguate: no visible binding for global variable ‘entity_name’
  disambiguate: no visible binding for global variable ‘..attr’
  disambiguate: no visible binding for global variable ‘hascompleteedge’
  disambiguate: no visible binding for global variable ‘edgeiscomplete’
  entity_consolidate_replicate: no visible binding for global variable
    ‘entity_name’
  find_acronyms: no visible binding for global variable ‘acronym’
  find_acronyms: no visible binding for global variable ‘name’
  find_acronyms: no visible binding for global variable ‘N’
  parse_text: no visible binding for global variable ‘eng_words’
  textnet_extract: no visible binding for global variable ‘keep’
  textnet_extract: no visible binding for global variable ‘dep_rel’
  textnet_extract: no visible binding for global variable ‘pos’
  textnet_extract: no visible global function definition for ‘.’
  textnet_extract: no visible binding for global variable ‘doc_id’
  textnet_extract: no visible binding for global variable ‘sentence_id’
  textnet_extract: no visible binding for global variable
    ‘source_or_target’
  textnet_extract: no visible binding for global variable ‘doc_sent_verb’
  textnet_extract: no visible binding for global variable ‘doc_sent’
  textnet_extract: no visible binding for global variable ‘head_verb_id’
  textnet_extract: no visible binding for global variable
    ‘doc_sent_parent’
  textnet_extract: no visible binding for global variable
    ‘parent_verb_id’
  textnet_extract: no visible binding for global variable
    ‘doc_sent_head_tok’
  textnet_extract: no visible binding for global variable ‘head_token_id’
  textnet_extract: no visible binding for global variable ‘neg’
  textnet_extract: no visible binding for global variable
    ‘head_verb_name’
  textnet_extract: no visible binding for global variable
    ‘head_verb_lemma’
  textnet_extract: no visible binding for global variable
    ‘head_verb_tense’
  textnet_extract: no visible binding for global variable ‘helper_lemma’
  textnet_extract: no visible binding for global variable ‘helper_token’
  textnet_extract: no visible binding for global variable ‘xcomp_verb’
  textnet_extract: no visible binding for global variable
    ‘xcomp_helper_lemma’
  textnet_extract: no visible binding for global variable
    ‘xcomp_helper_token’
  textnet_extract: no visible binding for global variable ‘has_sources’
  textnet_extract: no visible binding for global variable ‘entity_name’
  textnet_extract: no visible binding for global variable ‘entity_id’
  textnet_extract: no visible binding for global variable ‘entity_type’
  textnet_extract: no visible binding for global variable ‘row_id’
  textnet_extract: no visible binding for global variable ‘target’
  textnet_extract: no visible binding for global variable
    ‘hascompleteedge’
  textnet_extract: no visible binding for global variable
    ‘edgeiscomplete’
  textnet_extract: no visible binding for global variable ‘N’
  textnet_extract: no visible binding for global variable ‘num_mentions’
  textnet_extract: no visible binding for global variable
    ‘verb_classifications’
  top_features: no visible binding for global variable ‘fraction_of_doc’
  top_features: no visible binding for global variable
    ‘avg_fract_of_a_doc’
    . ..attr N acronym avg_fract_of_a_doc dep_rel doc_id doc_sent
    doc_sent_head_tok doc_sent_parent doc_sent_verb edgeiscomplete
```
