### 2025-11-04 #7615 @kbenoit
In the article, Zufall and Scott (2024) can be better cited:

> Zufall, Elise and Tyler A. Scott 2024. “ Syntactic Measurement of Governance Networks From Textual Data, With Application to Water Management Plans.” Policy Studies Journal 52(4): 941–954. https://doi.org/10.1111/psj.12556.
