### 2025-11-11 ucd-cepb/textnet#18 @ImBIOS
Add Community Guidelines (CONTRIBUTING.md and Issue Reporting Guidelines)
## Request from JOSS Review

This issue is part of the JOSS review process for textNet (see openjournals/joss-reviews#7615).

### Request

To improve community engagement and contribution guidelines, please add:

1. **CONTRIBUTING.md file** - Guidelines for how community members can contribute to the project
   - Code contribution process
   - Pull request guidelines
   - Code style/formatting standards
   - Testing requirements

2. **Issue reporting guidelines** - Clear instructions for reporting bugs and feature requests
   - Bug report template
   - Feature request template
   - Information to include when reporting issues

### Rationale

While not strictly required for JOSS publication, community guidelines help:
- Onboard new contributors
- Maintain code quality
- Streamline issue triage
- Encourage community participation

### Priority

**Minor revision** - This is a suggestion to improve the package, not a blocker for publication.

### Reference

- JOSS Review: https://github.com/openjournals/joss-reviews/issues/7615
