### 2025-11-11 ucd-cepb/textnet#19 @ImBIOS
Add Automated Test Suite (testthat) and CI/CD Workflow
## Request from JOSS Review

This issue is part of the JOSS review process for textNet (see openjournals/joss-reviews#7615).

### Request

To improve software quality and maintainability, please add:

1. **Automated test suite using testthat**
   - Unit tests for core functions (textnet_extract, parse_text, pdf_clean, etc.)
   - Integration tests for workflow components
   - Tests for edge cases and error handling
   - Note: The package currently has manual testing procedures in test_procedures.Rmd, which is good, but automated tests would be more robust

2. **CI/CD workflow for automated testing**
   - GitHub Actions workflow to run tests on pull requests
   - Tests on multiple R versions (if applicable)
   - Automated testing on different platforms (Linux, macOS, Windows)

### Rationale

Automated testing provides:
- Confidence in code changes
- Early detection of regressions
- Documentation of expected behavior
- Easier maintenance as the package grows

### Current State

The package includes:
- Manual testing procedures in `vignettes/test_procedures.Rmd`
- Good input validation in functions
- Example workflows in vignettes

Adding automated tests would complement these existing practices.

### Priority

**Minor revision** - This is a suggestion to improve the package, not a blocker for publication.

### Reference

- JOSS Review: https://github.com/openjournals/joss-reviews/issues/7615
- testthat documentation: https://testthat.r-lib.org/
- GitHub Actions for R: https://github.com/r-lib/actions
