### 2026-01-21 09f73a2
add trf model and entity ruler
- altered vignette location to satisfy build check warnings, altered parallel to 2 in vignettes to adhere to CRAN requirements
- added spacy trf model functionality, including GPU check/usage
- added NER specification using spacy entityruler
