### 2026-01-21 62413fb
pdftools suggested
suggested so that textnet is easier to install on remote computers, where I want to just transfer in processed files
