+++ DESCRIPTION
@@ -20,7 +20,7 @@ Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3
-Imports: 
+Imports:
     magrittr,
     dplyr,
     stringr,
@@ -31,7 +31,6 @@ Imports:
     data.table (>= 1.14.8),
     igraph,
     sna,
-    pdftools,
     ohenery,
     R.utils,
     methods,
@@ -41,7 +40,8 @@ Suggests:
     spacyr,
     rmarkdown (>= 2.21),
     knitr (>= 1.42),
-    ggraph
+    ggraph,
+    pdftools
 VignetteBuilder: knitr
 Depends: 
     R (>= 3.5.0)
