+++ NAMESPACE
@@ -51,8 +51,6 @@ importFrom(network,network.edgecount)
 importFrom(network,network.size)
 importFrom(ohenery,normalize)
 importFrom(pbapply,pblapply)
-importFrom(pdftools,pdf_ocr_text)
-importFrom(pdftools,pdf_text)
 importFrom(sna,centralization)
 importFrom(sna,connectedness)
 importFrom(sna,gtrans)
