### 2026-01-21 63a7c15
rename custom entity rule
spacy models alrady have an entity ruler, so have to have differnet name or else the custom one is ignored
