### 2026-01-22 f5b8433
claude helped fix gpu recogniction
