### 2026-01-22 62932a5
set prefer_gpu
