### 2026-01-22 5643ce9
adjust trycatch for trf load
