### 2026-01-22 dc860b8
gpu recognized but trf not working
