### 2026-01-23 387e01a
Update parse_text.R
