### 2026-01-23 2f8b0a5
set gpu allocator
