### 2026-01-23 e8d90a8
split out gpu version
