+++ man/parse_text_trf.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/parse_text.R
-\name{parse_text}
-\alias{parse_text}
-\title{Parse text using spaCy (CPU, en_core_web_lg model)}
+% Please edit documentation in R/parse_text_trf.R
+\name{parse_text_trf}
+\alias{parse_text_trf}
+\title{Parse text using spaCy transformer model with GPU support}
 \usage{
-parse_text(
+parse_text_trf(
   ret_path,
   keep_hyph_together = F,
   phrases_to_concatenate = NA,
@@ -49,6 +49,6 @@ The parsed data will be exported to these files.}
 A data.frame of tokens. For more information on the format, see the spacyr::spacy_parse help file
 }
 \description{
-Creates an edgelist and nodelist for each document using spaCy's en_core_web_lg model.
-For GPU-accelerated processing with the transformer model, use parse_text_trf().
+This function is specifically designed for the en_core_web_trf transformer model
+with GPU acceleration. For CPU-only processing with en_core_web_lg, use parse_text().
 }
