### 2026-01-23 e5158c4
make sure cuda gets located before reticulate
