### 2026-01-23 34418e7
trf version in python
skip spacyr
