+++ R/parse_text_trf.R
@@ -19,8 +19,9 @@
 #' @param entity_ruler_patterns A list of pattern dictionaries for spaCy's EntityRuler. Each pattern should be a named list with 'label' and 'pattern' elements. The 'pattern' can be a string for exact matches or a list of token attributes for complex patterns. If provided, these patterns will be added to spaCy's NLP pipeline. Use entity_specify() to create patterns from a dictionary.
 #' @param ruler_position A string controlling where EntityRuler is placed in the pipeline: "after" (default) places it after NER so custom patterns can override, "before" places it before NER so NER has final say.
 #' @param overwrite_ents A boolean. If TRUE (default), EntityRuler patterns override NER's entity assignments for overlapping spans. If FALSE, EntityRuler only fills gaps where NER found nothing.
-#' @return A data.frame of tokens. For more information on the format, see the spacyr::spacy_parse help file
-#' @importFrom data.table setDT
+#' @return A data.frame of tokens with columns: doc_id, sentence_id, token_id, token, lemma, pos, tag, head_token_id, dep_rel, entity.
+#'   Format is compatible with spacyr::spacy_parse output.
+#' @importFrom reticulate py_run_string py_eval py_config
 #' @importFrom stringr str_detect str_replace_all
 #' @importFrom stringi stri_replace_all_regex stri_escape_unicode
 #' @importFrom pbapply pblapply
@@ -31,8 +32,8 @@ parse_text_trf <- function(ret_path, keep_hyph_together=F, phrases_to_concatenat
                            concatenator="_", text_list, parsed_filenames,
                            overwrite=T, test=F, custom_entities = NULL, entity_ruler_patterns = NULL,
                            ruler_position = c("after", "before"), overwrite_ents = TRUE){
-  if(!requireNamespace("spacyr", quietly = T)){
-    stop("Package 'spacyr' must be installed to use this function.",
+  if(!requireNamespace("reticulate", quietly = T)){
+    stop("Package 'reticulate' must be installed to use this function.",
          call.=F)
   }
 
@@ -113,10 +114,6 @@ parse_text_trf <- function(ret_path, keep_hyph_together=F, phrases_to_concatenat
 
   # Set up Python environment
   Sys.setenv(RETICULATE_PYTHON=ret_path)
-  if(!requireNamespace("reticulate", quietly = T)){
-    stop("Package 'reticulate' must be installed to use this function.",
-         call.=F)
-  }
   reticulate::py_config()
 
   # Initialize GPU for spaCy transformer model
@@ -159,6 +156,11 @@ from thinc.api import get_current_ops
 ops = get_current_ops()
 assert ops is not None, 'get_current_ops() returned None'
 print(f'GPU initialized successfully. Ops type: {type(ops).__name__}')
+
+# Load the transformer model directly (bypassing spacyr::spacy_initialize which resets state)
+print('Loading en_core_web_trf model...')
+nlp = spacy.load('en_core_web_trf')
+print('Model loaded successfully!')
 ")
     TRUE
   }, error = function(e){
@@ -166,23 +168,72 @@ print(f'GPU initialized successfully. Ops type: {type(ops).__name__}')
                 "\n\nThis function requires GPU. For CPU processing, use parse_text() instead."))
   })
 
-  message("spaCy transformer configured to use GPU")
+  message("spaCy transformer model loaded with GPU")
 
-  # Initialize spaCy with transformer model
-  if(!stringr::str_detect("en_core_web_trf", "^en_")){
-    warning("This package was developed and tested on English texts.")
-  }
+  # Define Python parsing function (replaces spacyr::spacy_parse)
+  reticulate::py_run_string("
+def parse_texts(texts, doc_ids):
+    '''Parse texts using the loaded nlp model, returning spacyr-compatible format'''
[…]
