+++ NAMESPACE
@@ -52,6 +52,9 @@ importFrom(network,network.edgecount)
 importFrom(network,network.size)
 importFrom(ohenery,normalize)
 importFrom(pbapply,pblapply)
+importFrom(reticulate,py_config)
+importFrom(reticulate,py_eval)
+importFrom(reticulate,py_run_string)
 importFrom(sna,centralization)
 importFrom(sna,connectedness)
 importFrom(sna,gtrans)
