### 2026-01-23 ba490c4
Update parse_text_trf.R
