### 2026-01-23 #7615 @crvernon
:wave: @ezufall this review has been moving along for quite a long time.  We will need to wrap this up as soon as possible.  Can you give me a timeline for when you will be able to address the requested changes?

@amatsuo I see where you have [approved](https://github.com/openjournals/joss-reviews/issues/7615#issuecomment-3528862609) this submission for publication.  But I notice that you have several checkboxes in [your review checklist](https://github.com/openjournals/joss-reviews/issues/7615#issuecomment-2623265439) that still have not been checked off.  Could you please mark these as done if you believe it has been addressed?

@kbenoit and @ImBIOS please give a short update here for @ezufall of what is left outstanding that they need to address.

Thank you for your continued support and I hope we can free up your time as soon as possible.
