### 2026-01-23 1a1c692
cupy still not finding stuff
