### 2026-01-23 ef18cdb
change trf to run directly in python
