+++ R/parse_text_trf.R
@@ -6,7 +6,7 @@
 #' This function is specifically designed for the en_core_web_trf transformer model
 #' with GPU acceleration. For CPU-only processing with en_core_web_lg, use parse_text().
 #'
-#' @param ret_path filepath to use for Sys.setenv reticulate python call. Note: Python and miniconda must already be installed.
+#' @param python_path filepath to python executable with spacy, cupy, and en_core_web_trf installed.
 #' @param keep_hyph_together Set to true to replace hyphens within a single word with underscores. Defaults to false.
 #' @param phrases_to_concatenate character vector of phrases, in which each element is a string consisting of tokens separated by spaces. These are replaced with their concatenated version in order, from left to right. It is suggested that the most specific phrases, with the most words, are arranged at the left.
 #' @param concatenator This is a character or string that will be used to replace the spaces in the phrases_to_concatenate.
@@ -21,25 +21,29 @@
 #' @param overwrite_ents A boolean. If TRUE (default), EntityRuler patterns override NER's entity assignments for overlapping spans. If FALSE, EntityRuler only fills gaps where NER found nothing.
 #' @return A data.frame of tokens with columns: doc_id, sentence_id, token_id, token, lemma, pos, tag, head_token_id, dep_rel, entity.
 #'   Format is compatible with spacyr::spacy_parse output.
-#' @importFrom reticulate py_run_string py_eval py_config
+#' @importFrom jsonlite toJSON fromJSON
 #' @importFrom stringr str_detect str_replace_all
 #' @importFrom stringi stri_replace_all_regex stri_escape_unicode
 #' @importFrom pbapply pblapply
 #' @importFrom utils data
 #' @export
 
-parse_text_trf <- function(ret_path, keep_hyph_together=F, phrases_to_concatenate=NA,
+parse_text_trf <- function(python_path, keep_hyph_together=F, phrases_to_concatenate=NA,
                            concatenator="_", text_list, parsed_filenames,
                            overwrite=T, test=F, custom_entities = NULL, entity_ruler_patterns = NULL,
                            ruler_position = c("after", "before"), overwrite_ents = TRUE){
-  if(!requireNamespace("reticulate", quietly = T)){
-    stop("Package 'reticulate' must be installed to use this function.",
-         call.=F)
+
+  if(!requireNamespace("jsonlite", quietly = T)){
+    stop("Package 'jsonlite' must be installed to use this function.", call.=F)
   }
 
   # Input validation
-  if(!is.character(ret_path) || length(ret_path) != 1) {
-    stop("'ret_path' must be a single character string")
+  if(!is.character(python_path) || length(python_path) != 1) {
+    stop("'python_path' must be a single character string")
+  }
+
+  if(!file.exists(python_path)) {
+    stop(paste0("Python executable not found: ", python_path))
   }
 
   if(!is.logical(keep_hyph_together) || length(keep_hyph_together) != 1) {
@@ -95,185 +99,15 @@ parse_text_trf <- function(ret_path, keep_hyph_together=F, phrases_to_concatenat
     }
   }
 
-  # Fix OpenMP library conflict (common on macOS with conda)
-  Sys.setenv(KMP_DUPLICATE_LIB_OK = "TRUE")
-
-  # Prevent OpenMP thread conflicts between R packages (e.g., data.table) and PyTorch
-  Sys.setenv(OMP_NUM_THREADS = "1")
-  message("Note: Setting OMP_NUM_THREADS=1 to prevent OpenMP conflicts with transformer model.")
-
[…]
