### 2026-01-23 5fe4c9c
install rebuild with python trf
