+++ NAMESPACE
@@ -44,6 +44,8 @@ importFrom(igraph,simplify)
 importFrom(igraph,strength)
 importFrom(igraph,tail_of)
 importFrom(igraph,vertex_attr)
+importFrom(jsonlite,fromJSON)
+importFrom(jsonlite,toJSON)
 importFrom(magrittr,"%>%")
 importFrom(methods,is)
 importFrom(network,network)
@@ -52,9 +54,6 @@ importFrom(network,network.edgecount)
 importFrom(network,network.size)
 importFrom(ohenery,normalize)
 importFrom(pbapply,pblapply)
-importFrom(reticulate,py_config)
-importFrom(reticulate,py_eval)
-importFrom(reticulate,py_run_string)
 importFrom(sna,centralization)
 importFrom(sna,connectedness)
 importFrom(sna,gtrans)
