+++ R/zzz.R
@@ -1,0 +1,2 @@
+### package level settings ###
+.datatable.aware <- TRUE
