### 2026-01-29 bb2eed5
fully passed devtools::check
