+++ R/parse_text.R
@@ -165,13 +165,13 @@ parse_text <- function(ret_path, keep_hyph_together=F, phrases_to_concatenate=NA
     pages <- pbapply::pblapply(1:length(pages), function(i){
       stringi::stri_replace_all_regex(pages[i], pattern = phrases_to_concatenate,
                                       replacement = phrases_grouped,
-                                      vectorize= F)
+                                      vectorize_all = FALSE)
     })
   }
 
   if(keep_hyph_together){
     pages <- pbapply::pblapply(1:length(pages), function(i){
-      stringi::stri_replace_all_regex(pages[i], pattern= "(?<=\\w)[\\-\\u2013](?=\\w)", replacement ="_", vectorize=F)
+      stringi::stri_replace_all_regex(pages[i], pattern= "(?<=\\w)[\\-\\u2013](?=\\w)", replacement ="_", vectorize_all = FALSE)
     })
   }
 
