+++ vignettes/textNet_vignette_2025.Rmd
@@ -12,7 +12,7 @@ vignette: >
 
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
-knitr::opts_knit$set(root.dir = "~/Documents")
+#knitr::opts_knit$set(root.dir = "~/Documents")
 ```
 
 ## Introduction
