### 2026-01-29 b35b4e0
switch to pak package install
