### 2026-01-29 5d8285d
switch to pak package install (#22)
* switch to pak package install

* (auto) Paper PDF Draft

---------

Co-authored-by: tylerandrewscott <tylerandrewscott@users.noreply.github.com>
