+++ vignettes/paper.bib
@@ -76,6 +76,18 @@
 	date = {2022-10-11},
 }
 
+@software{csardi_pak_2024,
+	title = {pak: Another Approach to Package Installation},
+	rights = {{GPL}-3},
+	url = {https://cran.r-project.org/web/packages/pak/index.html},
+	shorttitle = {pak},
+	abstract = {The goal of pak is to make package installation faster and more reliable. In particular, it performs all HTTP operations in parallel, so metadata resolution and package downloads are fast. Metadata and package files are cached on the local disk as well.},
+	version = {0.8.0},
+	author = {Csárdi, Gábor and Hester, Jim and Software, Posit},
+	urldate = {2025-01-29},
+	date = {2024-09-20},
+}
+
 @software{csardi_igraph_2024,
 	title = {igraph: Network Analysis and Visualization},
 	rights = {{GPL}-2 {\textbar} {GPL}-3 [expanded from: {GPL} (≥ 2)]},
