### 2026-01-29 91b8c06
work in progress denotation
