### 2026-01-29 42aa663
work in progress denotation (#23)
