### 2026-02-02 0e0405a
initialize in python instead of spacyr
