+++ R/parse_text_trf.R
@@ -1,29 +1,24 @@
-# Exported function
-#parse_text_trf
-
-#' Parse text using spaCy transformer model with GPU support
+#' Parse text using spaCy transformer model
 #'
-#' This function is specifically designed for the en_core_web_trf transformer model
-#' with GPU acceleration. For CPU-only processing with en_core_web_lg, use parse_text().
-
-#' @description
-#' This function is a work in progress and should not be used yet.
-#' @keywords internal
+#' This function parses text using spaCy's en_core_web_trf transformer model.
+#' Unlike the CPU version (parse_text()), this function uses a pure Python
+#' implementation that allows PyTorch to handle GPU acceleration automatically.
+#'
 #' @param ret_path filepath to use for Sys.setenv reticulate python call. Note: Python and miniconda must already be installed.
 #' @param keep_hyph_together Set to true to replace hyphens within a single word with underscores. Defaults to false.
 #' @param phrases_to_concatenate character vector of phrases, in which each element is a string consisting of tokens separated by spaces. These are replaced with their concatenated version in order, from left to right. It is suggested that the most specific phrases, with the most words, are arranged at the left.
 #' @param concatenator This is a character or string that will be used to replace the spaces in the phrases_to_concatenate.
 #' @param text_list This is a named list, an object of the type resulting from pdf_clean, in which each list element is a document, and each string within a list element represents the text on one page
 #' @param parsed_filenames This is a character vector in which each element represents a filepath associated with its respective document.
-#' The parsed data will be exported to these files.
+#' The parsed data will be exported to these files (as .parquet files for efficiency).
 #' @param overwrite A boolean. Whether to overwrite existing files
 #' @param test A boolean. If TRUE and overwrite is FALSE, will still run parsing but won't save results. Useful for testing parsing on a subset without affecting saved files.
 #' @param custom_entities A named list. This does not overwrite the entity determination of the NLP engine, but rather catches user-defined entities that are not otherwise detected by the engine. Best used in combination with phrases_to_concatenate, since the custom entity label will only be applied if the entire token matches the definition. Does not search multiple consecutive tokens to define a match. These will be applied to all documents.
 #' @param entity_ruler_patterns A list of pattern dictionaries for spaCy's EntityRuler. Each pattern should be a named list with 'label' and 'pattern' elements. The 'pattern' can be a string for exact matches or a list of token attributes for complex patterns. If provided, these patterns will be added to spaCy's NLP pipeline. Use entity_specify() to create patterns from a dictionary.
 #' @param ruler_position A string controlling where EntityRuler is placed in the pipeline: "after" (default) places it after NER so custom patterns can override, "before" places it before NER so NER has final say.
 #' @param overwrite_ents A boolean. If TRUE (default), EntityRuler patterns override NER's entity assignments for overlapping spans. If FALSE, EntityRuler only fills gaps where NER found nothing.
-#' @return A data.frame of tokens. For more information on the format, see the spacyr::spacy_parse help file
-#' @importFrom data.table setDT
[…]
