+++ inst/python/parse_text_trf.py
@@ -1,0 +1,299 @@
+"""
+parse_text_trf.py - Pure Python spaCy parsing with transformer model
+
+This module provides text parsing using spaCy's en_core_web_trf transformer model.
+It is designed to be called from R via reticulate, bypassing spacyr to avoid
+GPU initialization issues.
+
+Output format matches spacyr::spacy_parse() for compatibility with existing textNet code.
+"""
+
+import os
+import json
+import spacy
+import pandas as pd
+from typing import List, Dict, Optional, Any
+from pathlib import Path
+
+
+class SpacyParserTRF:
+    """SpaCy parser using transformer model with optional EntityRuler support."""
+
+    def __init__(self, model: str = "en_core_web_trf"):
+        """
+        Initialize the parser with the specified model.
+
+        Args:
+            model: spaCy model name (default: en_core_web_trf)
+        """
+        self.model_name = model
+        self.nlp = None
+        self._initialized = False
+
+    def initialize(
+        self,
+        entity_ruler_patterns: Optional[List[Dict]] = None,
+        ruler_position: str = "after",
+        overwrite_ents: bool = True
+    ) -> None:
+        """
+        Load the spaCy model and configure pipeline.
+
+        Args:
+            entity_ruler_patterns: List of pattern dicts with 'label' and 'pattern' keys
+            ruler_position: "after" or "before" NER in pipeline
+            overwrite_ents: Whether EntityRuler overrides NER assignments
+        """
+        print(f"Loading spaCy model: {self.model_name}")
+        self.nlp = spacy.load(self.model_name)
+        print(f"Model loaded: {self.nlp.meta['name']}")
+
+        # Configure EntityRuler if patterns provided
+        if entity_ruler_patterns:
+            self._add_entity_ruler(entity_ruler_patterns, ruler_position, overwrite_ents)
+
+        self._initialized = True
+
+    def _add_entity_ruler(
+        self,
+        patterns: List[Dict],
+        position: str,
+        overwrite_ents: bool
+    ) -> None:
+        """Add EntityRuler to the pipeline."""
+        config = {"overwrite_ents": overwrite_ents}
+
+        if position == "after":
+            ruler = self.nlp.add_pipe("entity_ruler", after="ner", config=config)
+        else:
+            ruler = self.nlp.add_pipe("entity_ruler", before="ner", config=config)
+
+        ruler.add_patterns(patterns)
+        print(f"EntityRuler configured: position={position}, overwrite_ents={overwrite_ents}, {len(patterns)} patterns")
+
+    def parse_texts(
+        self,
+        texts: List[str],
+        doc_ids: List[str],
+        batch_size: int = 50
+    ) -> pd.DataFrame:
+        """
+        Parse a list of texts and return results in spacyr-compatible format.
+
+        Args:
+            texts: List of text strings to parse
+            doc_ids: List of document IDs corresponding to each text
+            batch_size: Batch size for nlp.pipe()
+
+        Returns:
+            DataFrame with columns: doc_id, sentence_id, token_id, token, lemma,
+                                   pos, tag, entity, head_token_id, dep_rel
+        """
+        if not self._initialized:
+            raise RuntimeError("Parser not initialized. Call initialize() first.")
+
+        if len(texts) != len(doc_ids):
+            raise ValueError("texts and doc_ids must have same length")
+
+        all_rows = []
+
+        # Process with nlp.pipe for efficiency
+        docs = list(self.nlp.pipe(texts, batch_size=batch_size))
+
+        for doc, doc_id in zip(docs, doc_ids):
+            doc_rows = self._parse_single_doc(doc, doc_id)
+            all_rows.extend(doc_rows)
+
+        df = pd.DataFrame(all_rows)
+
[…]
