+++ DESCRIPTION
@@ -38,6 +38,7 @@ Imports:
 Suggests:
     reticulate,
     spacyr,
+    arrow,
     rmarkdown (>= 2.21),
     knitr (>= 1.42),
     ggraph,
