+++ man/read_parsed_trf.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/parse_text_trf.R
+\name{read_parsed_trf}
+\alias{read_parsed_trf}
+\title{Read parsed text from Parquet file}
+\usage{
+read_parsed_trf(filepath)
+}
+\arguments{
+\item{filepath}{Path to the .parquet file}
+}
+\value{
+A data.frame with parsed tokens
+}
+\description{
+Helper function to read Parquet files created by parse_text_trf().
+}
