+++ NAMESPACE
@@ -10,6 +10,7 @@ export(find_acronyms)
 export(parse_text)
 export(parse_text_trf)
 export(pdf_clean)
+export(read_parsed_trf)
 export(suggest_entities)
 export(textnet_extract)
 export(top_features)
