### 2026-02-03 d4c70be
switch from arrow to duckdb
