+++ .github/issue_templates/feature_request.md
@@ -1,0 +1,10 @@
+---
+name: Feature request
+about: Feature request template
+title: ''
+labels: 'enhancement'
+---
+
+**Feature description**
+
+Please include a clear, complete description of the requested feature. Include a step-by-step explanation of the enhancement and explain its utility for the general user.
