+++ .github/issue_templates/bug_report.md
@@ -1,0 +1,20 @@
+---
+name: Bug report
+about: Bug report template
+title: ''
+labels: 'bug'
+---
+
+**Bug description**
+
+A clear and concise description of the bug and how it occurs. Explain how the bug deviates from the expected behavior.
+
+
+**Reproducible example**
+
+Provide a minimal reproducible example, including step-by-step instructions for someone else to create the bug.
+
+
+**Additional context**
+
+Add any other necessary context about the problem.
