+++ README.md
@@ -55,12 +55,8 @@ spacy_install()
 spacy_download_langmodel('en_core_web_lg')
 ```
 
-# Working on this package
-
-Make changes to the code, then run `devtools::load_all()` and test them.
-To update documentation and NAMESPACE file `devtools::document()`.
-To reinstall the package: `pak::local_install()`.
-Note that the vignettes both end up slightly larger than `devtools::check()` prefers, so running as `devtools::check(build_args = "--compact-vignettes=both")` is preferred.
+# Guidelines for Contributors
+See CONTRIBUTING.md for guidelines for working on this package.
 
 # Contact
 
