### 2026-02-11 c3f7571
performance improvements to auto_headfoot_remove tool
