+++ R/pdf_clean.R
@@ -134,69 +134,230 @@ pdf_clean <- function(pdfs, keep_pages=NULL, ocr=F, maxchar=10000, export_paths=
     
     if(auto_headfoot_remove == T){
       for(pagenum in 1:length(texts)){
+        #uses the first element because only splitting one string
+        #equivalent to str_split_1
         linebreaks <- str_split(texts[pagenum],"\\n")[[1]]
         
         #HEADER
+        #if more than six lines of text on the page, remove all title-case lines
+        #before whichever comes first: a non-title case line, or
+        #the last empty row on or before the first six rows of the page
         if(length(linebreaks)>6){
-          #search for last set of at least two \\n in a row 
-          #which corresponds to a row of linebreaks that has only \\s* in it
-          #before the sixth \\n 
-          #which is row six in linebreaks
-          #and delete it and all the rows above it
+          #search first six rows
           linebreakhead <- linebreaks[1:6]
+          #search for sets of at least two \\n in a row
+          #which correspond to a row of linebreaks that only has \\s* in it
           emptylines <- which(str_detect(linebreakhead,"^\\s*$"))
-          if(length(emptylines)>=1){
-            headercut <- emptylines[length(emptylines)]
-            linebreaks <- linebreaks[(headercut+1):length(linebreaks)]
-          }
-          #if there are no empty lines, don't remove anything
-        }else{
-          emptylines <- which(str_detect(linebreaks,"^\\s*$"))
-          if(length(emptylines)>=1){
-            #just remove everything before the first set of two \\n,
-            headercut <- emptylines[1]
-            if(length(linebreaks)>headercut){
-              linebreaks <- linebreaks[(headercut+1):length(linebreaks)]
-            }else{
-              #if there are no lines of text after the cut, make the text NA
-              linebreaks <- NA
+          
+          emptylinegroups <- sapply(c(1:length(emptylines)),
+                                    function(i) {ifelse(i!=1 && (emptylines[i]==emptylines[i-1]+1),NA,emptylines[i])
+                                    })
+          emptylinegroups <- emptylinegroups[!is.na(emptylinegroups)]
+          
+          #find last empty line if it exists and check everything before that for removal
+          #we got rid of na's, so this outside condition makes sure the var isn't empty
+          
+          if(length(emptylinegroups)> 0){
+            #this condition makes sure the last empty line isn't on line one
+            if(emptylinegroups[length(emptylinegroups)] > 1){
+              maxheaderlength <- emptylinegroups[length(emptylinegroups)] - 1
+              #checker to see if a previous line was preserved, which means all 
+              #potential header lines afterward should
+              #also be preserved.
+              prevlinepreserved = F
+              linegetspreserved <- vector(mode = "logical", length = maxheaderlength)
+              for(j in 1:maxheaderlength){
+                #if there are over six lowercase words, it's probably a sentence.
+                linegetspreserved[j] <- str_count(linebreaks[j], "\\s[a-z]") > 6 | 
+                  #alternatively, if there's at least four long lowercase words, it's probably a sentence
+                  #and may have a lot of capitalized entities in it
+                  str_count(linebreaks[j], "\\s[a-z]{7,}") > 3
+                if(linegetspreserved[j]==T){
+                  prevlinepreserved = T
+                  if(j < maxheaderlength){
+                    #break out of the loop and preserve later header lines if you find a 
+                    #sentence
+                    linegetspreserved[j:maxheaderlength] <- T
+                    break()
+                  }
+                }
+              }
+              #remove the parts of the text that aren't supposed to get preserved
+              if(any(linegetspreserved==F)){
+                linebreaks <- linebreaks[-which(!linegetspreserved)]
+              }
             }
           }else{
[…]
