### 2026-02-11 3f0e448
increment version number: clean build; auto_headfoot_remove improved
