### 2026-02-11 9bb518b
Add table parsing functionality to find_acronyms
