+++ R/find_acronyms.R
@@ -1,9 +1,10 @@
 # Exported function
 # pdf_clean
 
-#' Take a character vector and detect all parenthetical statements in which an acronym is defined within the character vector. 
+#' Take a character vector and detect all parenthetical statements in which an acronym is uniquely defined within the character vector. 
 #'
 #' @param str A character vector
+#' @param table_text Optional; used to detect acronyms from tables. A character vector representing the contents of a unique text file, such as the output of pdftools::pdf_text. Each element of the vector corresponds to a page of a pdf. It is recommended to use raw text output, rather than the results of textNet::pdf_clean, as the cleaning process removes white spaces that could hinder the detection of acronym tables. Single-letter acronyms are excluded. Parenthetically defined acronyms take precedence over table-defined acronyms. 
 #' 
 #' @import data.table
 #' @importFrom stringr str_split str_remove_all str_replace_all
@@ -15,7 +16,7 @@
 #' @export
 #' 
 
-find_acronyms <- function(str){
+find_acronyms <- function(str, table_text = NULL){
   # Input validation
   if(!is.character(str)) {
     stop("'str' must be a character vector")
@@ -71,5 +72,72 @@ find_acronyms <- function(str){
   acronym_matches <- acronym_matches[,N:=NULL]
   #reorder cols
   setcolorder(x=acronym_matches,neworder=c("name", "acronym"))
+  
+  if(!is.null(table_text)){
+    if(!is.character(table_text)){
+      stop("All elements in rawtexts must be character vectors.")
+    }
+    shorts <- vector(mode = "character", length = 0)
+    longs <- vector(mode = "character", length = 0)
+    mysplits <- strsplit(table_text, "\\n")
+    #for each page in project
+    for(j in 1:length(mysplits)){
+      #split on more than one space
+      linesplits <- strsplit(mysplits[[j]], "\\s\\s+")
+      #for each line in page
+      if(length(linesplits)>0){
+        for(k in 1:length(linesplits)){
+          #remove any empty strings
+          linesplits[[k]] <- linesplits[[k]][linesplits[[k]] != ""]
+          #if there's nothing else on that line, just those two strings:
+          if(length(linesplits[[k]]) == 2){
+            #if the capital letters match and one of them has lowercase and the other has uppercase
+            if(linesplits[[k]][1] == 
+               stringr::str_remove_all(linesplits[[k]][2], "[a-z]|_|\\s") & 
+               stringr::str_detect(linesplits[[k]][2], "[a-z]") &
+               stringr::str_detect(linesplits[[k]][1], "[A-Z]")){
+              shorts <- append(shorts, linesplits[[k]][1])
+              longs <- append(longs, linesplits[[k]][2])
+              #if the capital letters match and one of them has lowercase and the other has uppercase
+            }else if(linesplits[[k]][2] == 
+                     stringr::str_remove_all(linesplits[[k]][1], "[a-z]|_|\\s") &
+                     stringr::str_detect(linesplits[[k]][1], "[a-z]") &
+                     stringr::str_detect(linesplits[[k]][2], "[A-Z]")){
+              shorts <- append(shorts, linesplits[[k]][2])
+              longs <- append(longs, linesplits[[k]][1])
+            }
+          }
+        }
+      }
+      
+    }
+    if(length(longs)>0){
+      acron_table <- data.frame(name=character(length(longs)), acronym=character(length(longs)))
+      acron_table$name <- longs
+      acron_table$acronym <- shorts
+    }else{
+      acron_table <- data.frame(name = vector(mode = "character", length = 0),
+                                acronym = vector(mode = "character", length = 0))
+    }
+    
+    
+    
+    #get rid of one-letter acronyms
+    
[…]
