### 2026-02-12 d78f50f
testthat and minor paper edits
