+++ tests/testthat/test_vignette.R
@@ -1,0 +1,425 @@
+
+library(textNet)
+library(stringr)
+library(testthat)
+
+#download sample pdf data from state website
+#this is the same as vignettes/old.pdf and vignettes/new.pdf
+#that are included in the package, EXCEPT the package versions have pages that are maps or figures removed to shrink file size.
+
+#URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/2840"
+#download.file(URL, destfile = "vignettes/old.pdf", #method="curl")
+
+#URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/9625"
+#download.file(URL, destfile = "vignettes/new.pdf", method="curl")
+pdfs <- c(system.file("extdata", "old.pdf", package = "textNet"),         
+          system.file("extdata", "new.pdf", package = "textNet"))
+
+old_new_text <- textNet::pdf_clean(pdfs, ocr=F, maxchar=10000, 
+                                   export_paths=NULL, return_to_memory=T, suppressWarn = F, 
+                                   auto_headfoot_remove = T)
+names(old_new_text) <- c("old","new")
+
+#we expect one element per pdf
+test_that("number of elements equals number of files", {
+  expect_equal(length(old_new_text), length(pdfs))
+})
+
+
+### Pre-Processing Step II: Parse Text
+
+
+library(findpython)
+ret_path <- find_python_cmd(required_modules = c('spacy', 'en_core_web_lg'))
+
+water_bodies <- c("surface water", "Surface water", "groundwater", "Groundwater", 
+                  "San Joaquin River", "Cottonwood Creek", "Chowchilla Canal Bypass", 
+                  "Friant Dam", "Sack Dam", "Friant Canal", "Chowchilla Bypass", 
+                  "Fresno River", "Sacramento River", "Merced River","Chowchilla River", 
+                  "Bass Lake", "Crane Valley Dam", "Willow Creek", "Millerton Lake", 
+                  "Mammoth Pool", "Dam 6 Lake", "Delta","Tulare Lake", 
+                  "Madera-Chowchilla canal", "lower aquifer", "upper aquifer", 
+                  "upper and lower aquifers", "lower and upper aquifers", 
+                  "Lower aquifer", "Upper aquifer", "Upper and lower aquifers", 
+                  "Lower and upper aquifers")
+
+if(!requireNamespace("spacyr", quietly = T)){
+  stop("Package 'spacyr' must be installed to use this function.",
+       call.=F)
+}
+old_new_parsed <- textNet::parse_text(ret_path, 
+                                      keep_hyph_together = F, 
+                                      phrases_to_concatenate = water_bodies, 
+                                      concatenator = "_", 
+                                      text_list = old_new_text,   
+                                      parsed_filenames=c("old_parsed","new_parsed"), 
+                                      overwrite = T,
+                                      custom_entities = list(WATER = water_bodies))
+
+#expect all pages are preserved 
+for(k in 1:length(old_new_parsed)){
+  maxpage <- max(as.numeric(stringr::str_remove(old_new_parsed[[k]]$doc_id, "text")))
+  test_that("all pages are preserved", {
+    expect_equal(maxpage, length(old_new_text[[k]]))
+  })
+  
+}
+
+
+
+
+
+### Extraction Expectations
+
+  
+
+ent_types <- c('ORG','GPE','PERSON','WATER')
+extracts <- vector(mode="list",length=length(old_new_parsed))
+for(m in 1:length(old_new_parsed)){
+  extracts[[m]] <- textnet_extract(old_new_parsed[[m]],cl=2,
[…]
