+++ DESCRIPTION
@@ -46,7 +46,8 @@ Suggests:
     findpython,
     ggplot2,
     gridExtra,
-    testthat
+    testthat (>= 3.0.0)
 VignetteBuilder: knitr
 Depends: 
     R (>= 3.5.0)
+Config/testthat/edition: 3
