### 2026-02-12 7797b6a
minor paper edits
