+++ R/disambiguate.R
@@ -26,7 +26,7 @@
 #' The number of times the function is run is determined by the longest "chain" in which a value in 'to' is found in 'from', 
 #' which may in turn correspond to a 'to' value that is found in 'from', and so on. 
 #' @param concatenator The word boundary to look for when match_partial_entity is true. Defaults to "_". 
-#' @return a cleaned textnet extract. See textnet_extract help file for structure.
+#' @return a cleaned textNet extract. See textnet_extract help file for structure.
 #' 
 #' @import data.table
 #' @importFrom stringr str_detect str_remove_all str_replace_all
