### 2026-02-12 e241712
preserve attributes for statnet no isolates
