### 2026-02-13 caed74d
paper draft revisions
