+++ vignettes/paper.md
@@ -17,6 +17,29 @@ output: pdf_document
 
 ## Introduction
 
+The *textNet* package allows a user to input one or more
+PDF documents and create complex directed, multiplex, and
+multimodal network graphs. This enables rich analysis of the relationships between
+verb attributes and tenses, entity types, structural motifs, and other network 
+characteristics. Entities mentioned within the input text become nodes, and the 
+verbs connecting them in the sentences of the text become directed edges. [Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556)
+demonstrate the use of *textNet* to identify which actors are involved in start-up
+versus ongoing management tasks, characterize patterns of information 
+and funding flows, and compare the distribution of management tasks in networks from
+regions known to have contrasting characteristics [-@zufall_syntactic_2024]. 
+
+*textNet* has applications in social science research, including governance network scholarship, as demonstrated by
+[Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556) 
+[-@zufall_syntactic_2024] and by ongoing work on water resources governance and environmental impact assessments at the 
+UC Davis Center for Environmental Policy and Behavior. *textNet* also works on arbitrarily 
+long documents, making it well suited for research applications including legal 
+scholarship, social-ecological network analysis, 
+analysis of government planning documents, court proceedings, archival research, communication 
+and media research, and other fields interested in exploring events and entity 
+relationships in textual data.
+
+## Statement of Need
+
 Network measurement in social science typically relies on data collected
 through surveys and interviews. Document-based measurement can be automated and scaled, providing opportunities for large scale or
 longitudinal research that are not possible through traditional methods.
@@ -26,38 +49,16 @@ words within documents (such as the
 app [@levallois_translating_2012], the
 [“textnets”](https://github.com/cbail/textnets) package [@bail_cbailtextnets_2024],
 [InfraNodus](https://infranodus.com/) [@paranyushkin_infranodus_2018], and many
-more). But there is, to our knowledge, no open-source tool that
-generates network data based on the syntactic relationships between
-entities within a sentence. *textNet* allows a user to input one or more
-PDF documents and create arbitrarily complex directed, multiplex, and
-multimodal network graphs, enabling rich analysis of the relationships between
-verb attributes and tenses, entity types, structural motifs, and other network 
-characteristics. For instance, [Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556)
-demonstrate the use of *textNet* to identify which actors are involved in start-up
-versus ongoing management tasks, characterize patterns of information 
-and funding flows, and compare the distribution of management tasks in networks from
-regions known to have contrasting characteristics [-@zufall_syntactic_2024]. 
-*textNet* also works on arbitrarily long documents, making it well suited 
-for research applications using long texts such as government planning documents, 
-court proceedings, regulatory impact analyses, and environmental impact assessments.
-
-*textNet* has applications in governance network scholarship, as demonstrated by
-[Zufall and Scott](https://onlinelibrary.wiley.com/doi/full/10.1111/psj.12556) 
-[-@zufall_syntactic_2024] and by ongoing work on water resources governance at the 
-UC Davis Center for Environmental Policy and Behavior. Additional potential 
-applications include legal scholarship, social-ecological network analysis, 
-government planning documents, court proceedings, archival research, communication 
-and media research, and other fields interested in exploring events and entity 
-relationships in textual data.
-
-## Statement of Need
-
-Network extraction from the kinds of documents listed above has typically required manual coding.
[…]
