+++ vignettes/paper.bib
@@ -32,16 +32,17 @@
 
 @article{zufall_syntactic_2024,
 	title = {Syntactic measurement of governance networks from textual data, with application to water management plans},
-	volume = {n/a},
+	volume = {52},
 	rights = {© 2024 The Author(s). Policy Studies Journal published by Wiley Periodicals {LLC} on behalf of Policy Studies Organization.},
 	issn = {1541-0072},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/psj.12556},
 	doi = {10.1111/psj.12556},
 	abstract = {This paper demonstrates an automated workflow for extracting network data from policy documents. We use natural language processing tools, part-of-speech tagging, and syntactic dependency parsing, to represent relationships between real-world entities based on how they are described in text. Using a corpus of regional groundwater management plans, we demonstrate unique graph motifs created through parsing syntactic relationships and how document-level syntax can be aggregated to develop large-scale graphs. This approach complements and extends existing methods in public management and governance research by (1) expanding the feasible geographic and temporal scope of data collection and (2) allowing for customized representations of governance systems to fit different research applications, particularly by creating graphs with many different node and edge types. We conclude by reflecting on the challenges, limitations, and future directions of automated, text-based methods for governance research.},
-	issue = {n/a},
+	issue = {941--954},
+	number = {4},
 	journaltitle = {Policy Studies Journal},
 	author = {Zufall, Elise and Scott, Tyler A.},
-	urldate = {2024-10-25},
+	urldate = {2024-11-18},
 	date = {2024-07-01},
 	langid = {english},
 	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/psj.12556},
