### 2026-02-14 ad4371b
Update paper.md
