### 2026-02-14 #7615 @ezufall
Many thanks to all reviewers for the thorough feedback, which has greatly improved the package. Please see a summary below:

Revisions from ImBIOS:
1. The grammatical issues, clarity issues, content organization, and redundancy in the paper introduction vs. statement of need has been addressed.
2. Consistency issues regarding package name and function name formatting have been addressed
3. Installation instructions are moved after the “Overview and Main Functions” section
4. The testing procedures are now included in a testthat - compatible folder for automated testing.
5. We have maintained our current workflow for this version of the software. We hope to incorporate automated CI/CD workflow in a future update.

Revisions from kbenoit:
1. Installation procedures are updated to pak.
2. Build and install process is now clean.
3. Redundant license files are now removed.
4. The citation is now updated in the paper.

Revisions from amatsuo:
	N/A; these were addressed in the previous round of revisions.

Some minor performance improvements have been made to optional tools, particularly export_to_network(), and pdf_clean().
