+++ src/RcppExports.cpp
@@ -1,0 +1,39 @@
+// Generated by using Rcpp::compileAttributes() -> do not edit by hand
+// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+#include <Rcpp.h>
+
+using namespace Rcpp;
+
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
+// crawl_sentence_cpp
+List crawl_sentence_cpp(CharacterVector pos, CharacterVector dep_rel, IntegerVector head_token_id, CharacterVector lemma, CharacterVector token, IntegerVector token_id, CharacterVector tag);
+RcppExport SEXP _textNet_crawl_sentence_cpp(SEXP posSEXP, SEXP dep_relSEXP, SEXP head_token_idSEXP, SEXP lemmaSEXP, SEXP tokenSEXP, SEXP token_idSEXP, SEXP tagSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< CharacterVector >::type pos(posSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type dep_rel(dep_relSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type head_token_id(head_token_idSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type lemma(lemmaSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type token(tokenSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type token_id(token_idSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type tag(tagSEXP);
+    rcpp_result_gen = Rcpp::wrap(crawl_sentence_cpp(pos, dep_rel, head_token_id, lemma, token, token_id, tag));
+    return rcpp_result_gen;
+END_RCPP
+}
+
+static const R_CallMethodDef CallEntries[] = {
+    {"_textNet_crawl_sentence_cpp", (DL_FUNC) &_textNet_crawl_sentence_cpp, 7},
+    {NULL, NULL, 0}
+};
+
+RcppExport void R_init_textNet(DllInfo *dll) {
+    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
+    R_useDynamicSymbols(dll, FALSE);
+}
