+++ R/textnet_extract.R
@@ -66,6 +66,8 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
                                     return_to_memory = T, keep_incomplete_edges=F,
                                     remove_neg = T) {
 
+  doc_sent_parent.y <- NULL # silence R CMD check NOTE
+
   # Input validation
   if(!is.data.frame(x) && !is.data.table(x)) {
     stop("'x' must be a data.frame or data.table")
@@ -139,16 +141,16 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
   #in the non-matching tokens list and grab its entity_name; which is the full name of the appos. the 
   #abbrev is the entity_name of the appos group
   xappos <- x[x$source_or_target=="appos" & nchar(x$entity_name)>0 & x$entity_type%in%keep_entities,]
-  cat_splits <- split(xappos,list(xappos$doc_sent, xappos$entity_id),drop=T)
-  apposlist <- pbapply::pblapply(cat_splits, function(z) {
+  # Build a keyed lookup for x by doc_sent + token_id
+  data.table::setkeyv(x, c("doc_sent", "token_id"))
+  cat_splits <- split(xappos, list(xappos$doc_sent, xappos$entity_id), drop=T)
+  apposlist <- lapply(cat_splits, function(z) {
       anchor <- which(!(z$head_token_id %in% z$token_id))[1]
-      if(x[x$doc_sent==z$doc_sent[1] & x$token_id==z$head_token_id[anchor],"entity_type"]%in%keep_entities){
-        print(x[x$doc_sent==z$doc_sent[1] & x$token_id==z$head_token_id[anchor],])
-        as.data.table(cbind(unname(z[1,"entity_name"]),
-                            unname(x[x$doc_sent==z$doc_sent[1] & x$token_id==z$head_token_id[anchor],"entity_name"])))
-        
+      match_row <- x[.(z$doc_sent[1], z$head_token_id[anchor])]
+      if(nrow(match_row) > 0 && match_row$entity_type[1] %in% keep_entities){
+        data.table::data.table(V1 = z$entity_name[1], V2 = match_row$entity_name[1])
       }
-  }, cl=cl)
+  })
   apposlist <- data.table::rbindlist(apposlist)
   
   #if it's not empty, it will have two columns, which we can name.
@@ -202,23 +204,23 @@ textnet_extract <- function (x, concatenator = "_",file = NULL,cl = 1,
   
   sources <- source_target_list[has_sources==T & source_or_target=='source',]
     
-  ## find sources associated with the unsourced verbs' parent verbs
-  adopted_source_ids <- lapply(seq_along(unsourced_verbs$doc_sent_parent), function(i)
-           grep(unsourced_verbs$doc_sent_parent[i],sources$doc_sent_verb, value=F))
-  
-  #create a copy of the adopted sources, where the child doc_sent_verb 
-  #overwrites the original source's doc_sent_verb
-  adopted_sources <- lapply(seq_along(adopted_source_ids), function(i){
-      if(length(adopted_source_ids[[i]])>0){
-          tempsources <- sources[adopted_source_ids[[i]],]
-          tempsources$doc_sent_verb <- unsourced_verbs$doc_sent_verb[i]
-      }else{
-        tempsources <- NULL
-      }
-      tempsources
-  })
-    
-  adopted_sources_df <- data.table::rbindlist(adopted_sources, use.names=T,fill=T)
+  ## find sources associated with the unsourced verbs' parent verbs via merge
+  adopted_sources_df <- data.table::merge.data.table(
+    unsourced_verbs[, .(doc_sent_verb, doc_sent_parent)],
[…]
