+++ R/clean_entities.R
@@ -53,11 +53,9 @@ clean_entities <- function(v, remove_nums=T, remove_trailing_s=T, concatenator =
              "\\U0001d44c", "\\U0001d44d")
   letts <- c(letters,LETTERS)
   
-  v <- lapply(1:length(v), function(i){
-    stri_replace_all_regex(v[i], pattern = maths,
-                           replacement = letts,
-                           vectorize_all = FALSE)
-  })
+  v <- stri_replace_all_regex(v, pattern = maths,
+                              replacement = letts,
+                              vectorize_all = FALSE)
   
   if(remove_trailing_s==T){
     #remove strings with specific placement: trailing "'s"
