+++ R/parse_text.R
@@ -195,7 +195,7 @@ parse_text <- function(ret_path, keep_hyph_together=F, phrases_to_concatenate=NA
                                        nounphrase = T)
 
       lettertokens <- parsedtxt$token[stringr::str_detect(parsedtxt$token, "[a-zA-Z]")]
-      lettertokensunicodeescaped <- stringi::stri_escape_unicode(lettertokens)
+      lettertokensunicodeescaped <- stringi::stri_escape_unicode(tolower(lettertokens))
       utils::data(eng_words)
       pctlettersineng <- sum(lettertokensunicodeescaped %in% eng_words)/length(lettertokensunicodeescaped)
 
