+++ R/RcppExports.R
@@ -1,0 +1,7 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+crawl_sentence_cpp <- function(pos, dep_rel, head_token_id, lemma, token, token_id, tag) {
+    .Call('_textNet_crawl_sentence_cpp', PACKAGE = 'textNet', pos, dep_rel, head_token_id, lemma, token, token_id, tag)
+}
+
