### 2026-02-26 6c42c8a
fail if no-text files
