### 2026-03-04 2cbd85b
add sentence crawl testing suite
