+++ R/structural_org_patterns.R
@@ -1,0 +1,134 @@
+#' Build structural patterns for government organization names
+#'
+#' Creates spaCy EntityRuler patterns that match the grammatical shape of
+#' government organization names (e.g., "California Department of Fish &
+#' Wildlife") rather than enumerating exact strings. Designed to complement
+#' [entity_specify()], which matches exact dictionary entries.
+#'
+#' Patterns anchor on a set of organizational keywords (e.g., "department",
+#' "bureau") and require title-case tokens on either side. Three pattern
+#' templates are generated:
+#' 1. `{title+} ORG_WORD [of/for] {title+}` — plain multi-word suffix
+#' 2. `{title+} ORG_WORD [of/for] {title} and/& {title}` — coordinated suffix
+#' 3. `{title+} ORG_WORD of/for {title} {title} and/& {title}` — three-part
+#'    coordinated suffix
+#'
+#' `max_preceding` and `max_trailing` control how many title-case tokens are
+#' consumed on each side of the org keyword, letting you trade recall for
+#' precision.
+#'
+#' @param label A character string. The entity label assigned to all matched
+#'   spans. Defaults to `"PATTERN"`. Use a distinct label (e.g. `"PATTERN"`)
+#'   so structural matches can be distinguished from exact dictionary matches
+#'   produced by [entity_specify()].
+#' @param org_keywords A character vector of **lowercase** organizational
+#'   keywords used as anchors in patterns. Any token whose lowercase form
+#'   appears in this vector triggers a match attempt. Defaults to a built-in
+#'   set of common US government org terms.
+#' @param max_preceding Controls how many consecutive title-case tokens may
+#'   precede the organizational keyword. Accepts:
+#'   * A positive integer — caps the prefix length (e.g., `2L` covers
+#'     "California" and "San Bernardino"). Uses spaCy `OP = "{1,n}"`.
+#'   * `Inf` — unlimited; uses `OP = "+"`. Use with caution as this may
+#'     greedily consume long leading noun phrases.
+#'   * `TRUE` — alias for `Inf`. `FALSE` — alias for `1L`.
+#'
+#'   Defaults to `2L`, which covers the most common cases (one- and two-word
+#'   geographic prefixes such as "California" or "San Bernardino County").
+#' @param max_trailing Controls how many consecutive title-case tokens may
+#'   follow the organizational keyword phrase. Accepts the same values as
+#'   `max_preceding`. Defaults to `Inf` (unlimited), greedily capturing
+#'   multi-word suffixes such as "Board of Supervisors".
+#'
+#' @return A list of three pattern dictionaries formatted for spaCy's
+#'   EntityRuler, suitable for appending to the output of [entity_specify()]
+#'   and passing to [parse_text_trf()] via `entity_ruler_patterns`.
+#'
+#' @seealso [entity_specify()], [parse_text_trf()]
+#'
+#' @examples
+#' \dontrun{
+#' # Default: up to 2 preceding, unlimited trailing
+#' pats <- build_structural_org_patterns()
+#' str(pats)
+#'
+#' # Tighter matching: 2 preceding, at most 3 trailing
+#' pats2 <- build_structural_org_patterns(max_trailing = 3L)
+#'
+#' # Custom org keywords
+#' pats3 <- build_structural_org_patterns(
+#'   org_keywords = c("department", "bureau", "commission"),
+#'   max_preceding = 1L,
+#'   max_trailing  = 2L
+#' )
+#'
+#' # Combine with dictionary patterns from entity_specify()
+#' dict_pats <- entity_specify(my_names, entity_label = "DICT")
[…]
