+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(clean_entities)
 export(combine_networks)
 export(disambiguate)
+export(build_structural_org_patterns)
 export(entity_specify)
 export(export_to_network)
 export(filter_sentences)
