### 2026-03-05 c2944bd
installed update on macmini
