+++ man/build_structural_org_patterns.Rd
@@ -1,0 +1,95 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/structural_org_patterns.R
+\name{build_structural_org_patterns}
+\alias{build_structural_org_patterns}
+\title{Build structural patterns for government organization names}
+\usage{
+build_structural_org_patterns(
+  label = "PATTERN",
+  org_keywords = c("department", "bureau", "commission", "office", "board", "service",
+    "agency", "authority", "administration", "council", "division", "district",
+    "institute", "center"),
+  max_preceding = 2L,
+  max_trailing = Inf
+)
+}
+\arguments{
+\item{label}{A character string. The entity label assigned to all matched
+spans. Defaults to \code{"PATTERN"}. Use a distinct label (e.g. \code{"PATTERN"})
+so structural matches can be distinguished from exact dictionary matches
+produced by \code{\link[=entity_specify]{entity_specify()}}.}
+
+\item{org_keywords}{A character vector of \strong{lowercase} organizational
+keywords used as anchors in patterns. Any token whose lowercase form
+appears in this vector triggers a match attempt. Defaults to a built-in
+set of common US government org terms.}
+
+\item{max_preceding}{Controls how many consecutive title-case tokens may
+precede the organizational keyword. Accepts:
+\itemize{
+\item A positive integer — caps the prefix length (e.g., \code{2L} covers
+"California" and "San Bernardino"). Uses spaCy \code{OP = "{1,n}"}.
+\item \code{Inf} — unlimited; uses \code{OP = "+"}. Use with caution as this may
+greedily consume long leading noun phrases.
+\item \code{TRUE} — alias for \code{Inf}. \code{FALSE} — alias for \code{1L}.
+}
+
+Defaults to \code{2L}, which covers the most common cases (one- and two-word
+geographic prefixes such as "California" or "San Bernardino County").}
+
+\item{max_trailing}{Controls how many consecutive title-case tokens may
+follow the organizational keyword phrase. Accepts the same values as
+\code{max_preceding}. Defaults to \code{Inf} (unlimited), greedily capturing
+multi-word suffixes such as "Board of Supervisors".}
+}
+\value{
+A list of three pattern dictionaries formatted for spaCy's
+EntityRuler, suitable for appending to the output of \code{\link[=entity_specify]{entity_specify()}}
+and passing to \code{\link[=parse_text_trf]{parse_text_trf()}} via \code{entity_ruler_patterns}.
+}
+\description{
+Creates spaCy EntityRuler patterns that match the grammatical shape of
+government organization names (e.g., "California Department of Fish &
+Wildlife") rather than enumerating exact strings. Designed to complement
+\code{\link[=entity_specify]{entity_specify()}}, which matches exact dictionary entries.
+}
+\details{
+Patterns anchor on a set of organizational keywords (e.g., "department",
+"bureau") and require title-case tokens on either side. Three pattern
+templates are generated:
+\enumerate{
+\item \verb{\{title+\} ORG_WORD [of/for] \{title+\}} — plain multi-word suffix
+\item \verb{\{title+\} ORG_WORD [of/for] \{title\} and/& \{title\}} — coordinated suffix
+\item \verb{\{title+\} ORG_WORD of/for \{title\} \{title\} and/& \{title\}} — three-part
+coordinated suffix
+}
+
+\code{max_preceding} and \code{max_trailing} control how many title-case tokens are
+consumed on each side of the org keyword, letting you trade recall for
+precision.
+}
+\examples{
+\dontrun{
+# Default: up to 2 preceding, unlimited trailing
[…]
