+++ NAMESPACE
@@ -1,9 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+export(build_structural_org_patterns)
 export(clean_entities)
 export(combine_networks)
 export(disambiguate)
-export(build_structural_org_patterns)
 export(entity_specify)
 export(export_to_network)
 export(filter_sentences)
