### 2026-03-07 a6b11da
updating vignettes to html for 1.0.3
