### 2026-04-02 94bdef5
optional handling for sentences spanning multiple pages.
