+++ man/pdf_clean.Rd
@@ -18,7 +18,8 @@ pdf_clean(
   export_paths = NULL,
   return_to_memory = T,
   suppressWarn = F,
-  auto_headfoot_remove = T
+  auto_headfoot_remove = T,
+  sentence_shift = F
 )
 }
 \arguments{
@@ -47,6 +48,8 @@ the output text to memory. Defaults to T.}
 \item{suppressWarn}{Boolean; F by default. Setting this to T suppresses pdf_text conversion warnings, which can be verbose.}
 
 \item{auto_headfoot_remove}{Boolean. Tries to remove headers and footers if T. May remove small amounts of paragraph text, especially if there is no header on a page. Defaults to T to avoid having entities that appear consistently in headers be artificially inflated in the textnet_extract network. Recommend keeping this parameter T unless using a separate header/footer removal tool.}
+
+\item{sentence_shift}{Boolean. If T, tries to shift a beginning partial sentence to the previous page.}
 }
 \value{
 If return_to_memory is T, returns a list with a length equal
