### 2026-04-02 d8e85d5
test data update: pages with maps or figures removed to shrink file size
