+++ vignettes/textNet_vignette_2025.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "textNet: Directed, Multiplex, Multimodal Event Network Extraction from Textual Data"
 author: Elise Zufall and Tyler Scott
-date: Updated 6 March 2026
+date: Updated 2 Apr 2026
 bibliography: paper.bib
 output: rmarkdown::html_vignette
 vignette: >
@@ -293,7 +293,7 @@ The combine_networks function allows a composite network to be generated from mu
 
 For best results, composite network generation should not be done without an adequate disambiguation in Step 4. A function is included that merges the edgelists and nodelists of all documents. If the same node name is mentioned in multiple documents, the node attributes associated with the highest total number of edges for that node name are preserved.
 
-```{r composite}
+```{r composite, fig.width=5,fig.height=5, fig.retina=2}
 #we generate a composite network with the function below
 composite_net <- combine_networks(list(old_extract_net[[1]], new_extract_net[[1]]), 
                                      mode = "weighted")
@@ -425,7 +425,7 @@ with(combineddf,table(plan_version,num_graphs_in))
 
 We can also investigate differences in network statistics between the two plans. For instance, the distribution of degree does not change much between plan versions. The distribution of betweenness, likewise, is relatively stable except for person nodes, which are the least common nodes in the graph.
 
-```{r step8b4, warning=F, message=F}
+```{r step8b4, warning=F, message=F, fig.width=10,fig.height=8, fig.retina=2}
 library(gridExtra)
 library(ggplot2)
 #summary of differences between the two plans
