### 2026-04-03 #7615 @kbenoit
I tried to find my checklist above, but could not. I seem to remember not being able to get past building the package. That still seems to be the case: I cannot run the full check on the cloned repo. The instructions ought to be sufficient to do this. I have followed those on the repo home page and it still does not work on my macOS system.

```
E  creating vignettes (1m 13.3s)
   --- re-building ‘test_procedures.Rmd’ using rmarkdown

   Quitting from test_procedures.Rmd:49-85 [parse]
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   <error/rlang_error>
   Error in `find_python_cmd()`:
   ! Couldn't find a sufficient Python binary. If you haven't installed the Python dependency yet please do so. If you have but it isn't on the system path (as is default on Windows) please add it to path or set options('python_cmd'='/path/to/binary')  or set the PYTHON, PYTHON2, or PYTHON3 environmental variables.   Python must have access to the modules: spacy, en_core_web_lg
   ---
   Backtrace:
       ▆
    1. └─findpython::find_python_cmd(required_modules = c("spacy", "en_core_web_lg"))
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

   Error: processing vignette 'test_procedures.Rmd' failed with diagnostics:
   Couldn't find a sufficient Python binary. If you haven't installed the Python dependency yet please do so. If you have but it isn't on the system path (as is default on Windows) please add it to path or set options('python_cmd'='/path/to/binary')  or set the PYTHON, PYTHON2, or PYTHON3 environmental variables.   Python must have access to the modules: spacy, en_core_web_lg
   --- failed re-building ‘test_procedures.Rmd’

   --- re-building ‘textNet_vignette_2025.Rmd’ using rmarkdown

   Quitting from textNet_vignette_2025.Rmd:512-545 [parse]
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   <error/rlang_error>
   Error in `find_python_cmd()`:
   ! Couldn't find a sufficient Python binary. If you haven't installed the Python dependency yet please do so. If you have but it isn't on the system path (as is default on Windows) please add it to path or set options('python_cmd'='/path/to/binary')  or set the PYTHON, PYTHON2, or PYTHON3 environmental variables.   Python must have access to the modules: spacy, en_core_web_lg
   ---
   Backtrace:
       ▆
    1. └─findpython::find_python_cmd(required_modules = c("spacy", "en_core_web_lg"))
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

   Error: processing vignette 'textNet_vignette_2025.Rmd' failed with diagnostics:
   Couldn't find a sufficient Python binary. If you haven't installed the Python dependency yet please do so. If you have but it isn't on the system path (as is default on Windows) please add it to path or set options('python_cmd'='/path/to/binary')  or set the PYTHON, PYTHON2, or PYTHON3 environmental variables.   Python must have access to the modules: spacy, en_core_web_lg
   --- failed re-building ‘textNet_vignette_2025.Rmd’

   SUMMARY: processing the following files failed:
     ‘test_procedures.Rmd’ ‘textNet_vignette_2025.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error in `(function (command = NULL, args = character(), error_on_status = TRUE, …`:
! System command 'R' failed
---
Exit status: 1
stdout & stderr: <printed>
---
Backtrace:
 1. devtools::check(args = c("--as-cran"), build_args = c("--resave-data=best"), …
 2. withr::with_envvar(pkgbuild::compiler_flags(FALSE), action = "prefix", …
 3. base::force(code)
 4. pkgbuild::build(pkg$path, tempdir(), args = build_args, quiet = quiet, …
 5. pkgbuild:::withr_with_makevars(compiler_flags(debug = FALSE), { …
 6. pkgbuild:::withr_with_envvar(c(R_MAKEVARS_USER = makevars_file), { …
 7. base::force(code)
 8. base::force(code)
 9. pkgbuild:::withr_with_temp_libpaths(rcmd_build_tools(options$cmd, c(options$path, …
10. base::force(code)
11. pkgbuild::rcmd_build_tools(options$cmd, c(options$path, options$args), …
12. pkgbuild::with_build_tools({ …
13. base::withCallingHandlers(callr::rcmd_safe(..., env = env, spinner = FALSE, …
14. callr::rcmd_safe(..., env = env, spinner = FALSE, show = FALSE, …
15. callr:::run_r(options)
16. base::with(options, with_envvar(env, do.call(processx::run, c(list(bin, …
17. base::with.default(options, with_envvar(env, do.call(processx::run, …
18. base::eval(substitute(expr), data, enclos = parent.frame())
19. base::eval(substitute(expr), data, enclos = parent.frame())
20. callr:::with_envvar(env, do.call(processx::run, c(list(bin, args = real_cmdargs, …
21. base::force(code)
22. base::do.call(processx::run, c(list(bin, args = real_cmdargs, stdout_line_callback = real_callback(stdout), …
23. (function (command = NULL, args = character(), error_on_status = TRUE, …
24. base::throw(new_process_error(res, call = sys.call(), echo = echo, …
25. | base::signalCondition(cond)
26. (function (e) …
27. asNamespace("callr")$err$throw(e)
Execution halted

Exited with status 1.
```
