### 2026-04-05 92a5b37
new find python method
- also updated igraph names to remove deprecated names from old version
