+++ NAMESPACE
@@ -26,13 +26,13 @@ importFrom(dplyr,left_join)
 importFrom(dplyr,summarize)
 importFrom(igraph,E)
 importFrom(igraph,V)
-importFrom(igraph,as.undirected)
 importFrom(igraph,as_data_frame)
+importFrom(igraph,as_undirected)
 importFrom(igraph,cluster_louvain)
 importFrom(igraph,degree)
 importFrom(igraph,delete.edges)
-importFrom(igraph,delete.vertices)
 importFrom(igraph,delete_edge_attr)
+importFrom(igraph,delete_vertices)
 importFrom(igraph,edge_attr)
 importFrom(igraph,edge_attr_names)
 importFrom(igraph,graph_from_data_frame)
