### 2026-04-06 e4be4ce
package hygene
- test procedures out of vignette
- load packaged data don't re-parse
- vignette html not in tarball
