+++ vignettes/textNet_vignette_2025.Rmd
@@ -479,7 +479,7 @@ This appendix describes the pre-processing tools available through the *textNet*
 
 This is a wrapper for pdftools, which has the option of using pdf_text or OCR. We have also added an optional header/footer removal tool. This optional tool is solely based on carriage returns in the first or last few lines of the document, so may inadvertently remove portions of paragraphs. However, not removing headers or footers can lead to improper inclusion of header and footer material in sentences, artificially inflating the presence of nodes whose entity names are included in the header and footer. Because of the risk of headers and footers to preferentially inflate the presence of a few nodes, the header/footer remover is included by default. It can be turned off if the user has a preferred header/footer removal tool to use instead, or if the input documents lack headers and footers.
 
-```{r pdf_clean, eval = T}
+```{r pdf_clean, eval = FALSE}
 library(textNet)
 library(stringr)
 #download sample pdf data from state website
@@ -488,19 +488,27 @@ library(stringr)
 
 #URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/2840"
 #download.file(URL, destfile = "inst/extdata/old.pdf", method="curl")
-   
+
 #URL <- "https://sgma.water.ca.gov/portal/service/gspdocument/download/9625"
 #download.file(URL, destfile = "inst/extdata/new.pdf", method="curl")
- 
-pdfs <- c(system.file("extdata", "old.pdf", package = "textNet"),         
+
+pdfs <- c(system.file("extdata", "old.pdf", package = "textNet"),
             system.file("extdata", "new.pdf", package = "textNet"))
 
 #this is the equivalent of the textNet::old_new_text sample data
 #that travels with the package
-old_new_text <- textNet::pdf_clean(pdfs, ocr=F, maxchar=10000, 
-                     export_paths=NULL, return_to_memory=T, suppressWarn = F, 
+old_new_text <- textNet::pdf_clean(pdfs, ocr=F, maxchar=10000,
+                     export_paths=NULL, return_to_memory=T, suppressWarn = F,
                      auto_headfoot_remove = T)
 names(old_new_text) <- c("old","new")
+```
+
+The above code produces results equivalent to the bundled sample data, which can be loaded directly:
+
+```{r pdf_clean_load, echo = FALSE}
+library(textNet)
+library(stringr)
+old_new_text <- textNet::old_new_text
 ```
 
 ### Pre-Processing Step II: Parse Text
@@ -509,21 +517,21 @@ This is a wrapper for the pre-trained multipurpose NLP model *spaCy* [@honnibal_
 
[…]
