### 2026-04-07 06b27fc
remove old data shrink tar
- removed old data used for piloting
- created data-raw/ (buildignored) for helper functions creating the eng_words.rda and verb rda files
-
