+++ NAMESPACE
@@ -62,11 +62,14 @@ importFrom(stringi,stri_match_all)
 importFrom(stringi,stri_match_last)
 importFrom(stringi,stri_read_lines)
 importFrom(stringi,stri_replace_all_regex)
+importFrom(stringr,fixed)
+importFrom(stringr,str_count)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_remove)
 importFrom(stringr,str_remove_all)
 importFrom(stringr,str_replace_all)
 importFrom(stringr,str_split)
+importFrom(stringr,str_split_i)
 importFrom(tidyr,expand)
 importFrom(tidyr,tibble)
 importFrom(tools,file_ext)
