### 2026-04-07 4e91983
vignette re-org
separate /paper subdirectory
/paper and /vignettes have unique .bib files
old data cleanup
