+++ vignettes/vignette_references.bib
@@ -1,0 +1,267 @@
+
+@software{hollway_manynet_2024,
+	title = {manynet: Many Ways to  Make, Modify, Map, Mark, and Measure Myriad Networks},
+	url = {https://CRAN.R-project.org/package=manynet},
+	version = {1.2.6},
+	author = {Hollway, James},
+	date = {2024},
+}
+
+@article{levallois_translating_2012,
+	title = {Translating upwards: linking the neural and social sciences via neuroeconomics},
+	volume = {13},
+	url = {https://nocodefunctions.com/cowo/semantic_networks_tool.html},
+	pages = {789--797},
+	number = {11},
+	journaltitle = {Nature Reviews Neuroscience},
+	author = {Levallois, C. and Clithero, J. A. and Wouters, P. and Smidts, A. and Huettel, S. A.},
+	date = {2012},
+}
+
+@software{bail_cbailtextnets_2024,
+	title = {cbail/textnets},
+	rights = {{MIT}},
+	url = {https://github.com/cbail/textnets},
+	abstract = {R package to perform automated text analysis using network techniques},
+	version = {0.1.1},
+	author = {Bail, Chris},
+	urldate = {2024-10-24},
+	date = {2024-10-07},
+	note = {original-date: 2017-10-04T00:13:31Z},
+}
+
+@article{zufall_syntactic_2024,
+	title = {Syntactic measurement of governance networks from textual data, with application to water management plans},
+	volume = {52},
+	rights = {© 2024 The Author(s). Policy Studies Journal published by Wiley Periodicals {LLC} on behalf of Policy Studies Organization.},
+	issn = {1541-0072},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/psj.12556},
+	doi = {10.1111/psj.12556},
+	abstract = {This paper demonstrates an automated workflow for extracting network data from policy documents. We use natural language processing tools, part-of-speech tagging, and syntactic dependency parsing, to represent relationships between real-world entities based on how they are described in text. Using a corpus of regional groundwater management plans, we demonstrate unique graph motifs created through parsing syntactic relationships and how document-level syntax can be aggregated to develop large-scale graphs. This approach complements and extends existing methods in public management and governance research by (1) expanding the feasible geographic and temporal scope of data collection and (2) allowing for customized representations of governance systems to fit different research applications, particularly by creating graphs with many different node and edge types. We conclude by reflecting on the challenges, limitations, and future directions of automated, text-based methods for governance research.},
+	issue = {941--954},
+	number = {4},
+	journaltitle = {Policy Studies Journal},
+	author = {Zufall, Elise and Scott, Tyler A.},
+	urldate = {2024-11-18},
+	date = {2024-07-01},
+	langid = {english},
+	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/psj.12556},
+	file = {Full Text PDF:/Users/elisemiller/Zotero/storage/Y59XFXEI/Zufall and Scott - Syntactic measurement of governance networks from .pdf:application/pdf;Snapshot:/Users/elisemiller/Zotero/storage/W8IZZTGM/psj.html:text/html},
+}
+
+@software{paranyushkin_infranodus_2018,
+	title = {{InfraNodus}},
+	url = {https://infranodus.com/},
[…]
