+++ NAMESPACE
@@ -14,7 +14,6 @@ export(suggest_entities)
 export(textnet_extract)
 export(top_features)
 import(data.table)
-importFrom(R.utils,gunzip)
 importFrom(data.table,rbindlist)
 importFrom(data.table,setDT)
 importFrom(dplyr,arrange)
@@ -60,7 +59,6 @@ importFrom(stats,median)
 importFrom(stringi,stri_escape_unicode)
 importFrom(stringi,stri_match_all)
 importFrom(stringi,stri_match_last)
-importFrom(stringi,stri_read_lines)
 importFrom(stringi,stri_replace_all_regex)
 importFrom(stringr,fixed)
 importFrom(stringr,str_count)
@@ -72,13 +70,6 @@ importFrom(stringr,str_split)
 importFrom(stringr,str_split_i)
 importFrom(tidyr,expand)
 importFrom(tidyr,tibble)
-importFrom(tools,file_ext)
 importFrom(utils,data)
-importFrom(utils,download.file)
 importFrom(utils,read.csv)
 importFrom(utils,read.table)
-importFrom(utils,untar)
-importFrom(utils,unzip)
-importFrom(xml2,read_xml)
-importFrom(xml2,xml_attr)
-importFrom(xml2,xml_find_all)
