+++ .github/workflows/draft-pdf.yml
@@ -2,8 +2,8 @@ name: Draft PDF
 on: 
   push:
     paths:
-      - vignettes/paper.md
-      - vignettes/paper.bib
+      - paper/paper.md
+      - paper/paper.bib
       - .github/workflows/draft-pdf.yml
 
 jobs:
@@ -18,7 +18,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: vignettes/paper.md
+          paper-path: paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -26,10 +26,10 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: vignettes/paper.pdf
+          path: paper/paper.pdf
       - name: Commit PDF to repository
         uses: EndBug/add-and-commit@v9
         with:
           message: '(auto) Paper PDF Draft'
           # This should be the path to the paper within your repo.
-          add: 'vignettes/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
+          add: 'paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
