### 2026-04-07 a2d4f28
minor typo corrections + check and rebuild
