### 2026-04-08 dac7ed4
split vignette test
remove whole vignette test, instead have separate pdf_clean test and various extract/tag tests. skip the spacy parse test since that test is of spacy. tests now rely on bundled data, not data stored in test suite
