+++ tests/testthat/test-network.R
@@ -1,0 +1,170 @@
+
+library(textNet)
+library(testthat)
+library(stringr)
+
+# Use bundled data; no spaCy needed
+old_new_parsed <- textNet::old_new_parsed
+old_new_text   <- textNet::old_new_text
+
+# water_bodies and ent_types come from helper-fixtures.R
+
+# --- Setup: extraction + disambiguation (mirrors test-disambiguate.R setup) ---
+
+extracts <- lapply(old_new_parsed, function(parsed) {
+  textnet_extract(parsed, cl = 2,
+                  keep_entities = ent_types,
+                  keep_incomplete_edges = TRUE)
+})
+
+build_tofrom <- function(acronyms) {
+  data.table::data.table(
+    from = c(
+      as.list(acronyms$acronym),
+      list("Sub_basin", "Sub_Basin",
+           "upper_and_lower_aquifers", "Upper_and_lower_aquifers",
+           "Lower_and_upper_aquifers", "lower_and_upper_aquifers")
+    ),
+    to = c(
+      as.list(acronyms$name),
+      list("Subbasin", "Subbasin",
+           c("upper_aquifer", "lower_aquifer"),
+           c("upper_aquifer", "lower_aquifer"),
+           c("upper_aquifer", "lower_aquifer"),
+           c("upper_aquifer", "lower_aquifer"))
+    )
+  )
+}
+
+old_acronyms      <- find_acronyms(old_new_text[[1]])
+new_acronyms      <- find_acronyms(old_new_text[[2]])
+
+old_extract_clean <- disambiguate(
+  textnet_extract      = extracts[[1]],
+  from                 = build_tofrom(old_acronyms)$from,
+  to                   = build_tofrom(old_acronyms)$to,
+  match_partial_entity = c(rep(FALSE, nrow(old_acronyms)), TRUE, TRUE, FALSE, FALSE, FALSE, FALSE)
+)
+
+new_extract_clean <- disambiguate(
+  textnet_extract      = extracts[[2]],
+  from                 = build_tofrom(new_acronyms)$from,
+  to                   = build_tofrom(new_acronyms)$to,
+  match_partial_entity = c(rep(FALSE, nrow(new_acronyms)), TRUE, TRUE, FALSE, FALSE, FALSE, FALSE)
+)
+
+set.seed(50000)
+old_extract_net <- export_to_network(old_extract_clean, "igraph",
+                                     keep_isolates = FALSE,
+                                     collapse_edges = FALSE,
+                                     self_loops = TRUE)
+set.seed(50000)
+new_extract_net <- export_to_network(new_extract_clean, "igraph",
+                                     keep_isolates = FALSE,
+                                     collapse_edges = FALSE,
+                                     self_loops = TRUE)
+
+composite_net <- combine_networks(
+  list(old_extract_net[[1]], new_extract_net[[1]]),
+  mode = "weighted"
+)
+
+# --- export_to_network: return structure ---
+
+test_that("export_to_network returns a 2-element list for old network", {
+  expect_type(old_extract_net, "list")
+  expect_length(old_extract_net, 2)
+})
+
+test_that("old network first element is an igraph", {
+  expect_true(igraph::is_igraph(old_extract_net[[1]]))
+})
+
+test_that("new network first element is an igraph", {
+  expect_true(igraph::is_igraph(new_extract_net[[1]]))
+})
+
+test_that("network stats table has expected columns", {
+  expected_cols <- c("num_nodes", "num_edges", "connectedness",
+                     "centralization", "transitivity",
+                     "pct_entitytype_homophily", "reciprocity",
[…]
