+++ NAMESPACE
@@ -8,7 +8,6 @@ export(export_to_network)
 export(filter_sentences)
 export(find_acronyms)
 export(parse_text)
-export(parse_text_trf)
 export(pdf_clean)
 export(suggest_entities)
 export(textnet_extract)
