### 2026-04-08 a1dc427
replace T/F with TRUE/FALSE
