+++ R/filter_sentences.R
@@ -7,12 +7,12 @@
 #' All non-word characters are removed.
 #' Consecutive underscores are collapsed to a single underscore.
 #' Leading and trailing underscores are removed.
-#' Entities that have no letters are removed, if remove_nums is set to T.
+#' Entities that have no letters are removed, if remove_nums is set to TRUE.
 #'
 #' @param file Data frame to be cleaned. To ensure file format appropriateness, it should be of the form that results from spacy_parse output
 #' @param filter Dictionary on which to filter file for acceptable words. Defaults to SCOWL 2020.12.07, including files with extension <= 60, with filenames that include "variant", "american", "british", canadian", or "australian".
 #' @param percent_threshold The required percentage of tokens in a sentence required to be found in the filter to keep the sentence in the dataset. Defaults to 40.
-#' @param case_sensitive Whether the token is required to have the specific casing used in the dictionary to count as a match. Defaults to F.
+#' @param case_sensitive Whether the token is required to have the specific casing used in the dictionary to count as a match. Defaults to FALSE.
 #' @return a cleaned version of 'file', keeping only the sentences that pass the threshold requirement.
 #' @importFrom stats aggregate
 #' @importFrom dplyr left_join
@@ -21,7 +21,7 @@
 #' 
 
 filter_sentences <- function(file, filter = textNet::eng_words, 
-                             percent_threshold = 40, case_sensitive = F){
+                             percent_threshold = 40, case_sensitive = FALSE){
   # Input validation
   if(!is.data.frame(file)) {
     stop("'file' must be a data frame")
