### 2026-04-08 d9e3daf
lazy load data out of loop
also call data from textnet instead of with data() load from utils
