### 2026-04-08 46fa883
fixes from code review
- switched up how the dplyr + global variables thign is handled. where possible, use .data$var in dply calls
- added dplyr:: to things like dplyr::filter to avoid overlap with stats::filter
- add unlist() to lapply() in clean_entities becuase it wants to pass vector
