+++ NAMESPACE
@@ -44,6 +44,7 @@ importFrom(igraph,tail_of)
 importFrom(igraph,vertex_attr)
 importFrom(magrittr,"%>%")
 importFrom(methods,is)
+importFrom(rlang,.data)
 importFrom(network,network)
 importFrom(network,network.density)
 importFrom(network,network.edgecount)
@@ -69,6 +70,5 @@ importFrom(stringr,str_split)
 importFrom(stringr,str_split_i)
 importFrom(tidyr,expand)
 importFrom(tidyr,tibble)
-importFrom(utils,data)
 importFrom(utils,read.csv)
 importFrom(utils,read.table)

[…]
