+++ man/clean_entities.Rd
@@ -8,16 +8,21 @@ Any trailing "'s" at the end of the entity name is removed.
 All non-word characters are removed.
 Consecutive underscores are collapsed to a single underscore.
 Leading and trailing underscores are removed.
-Entities that have no letters are removed, if remove_nums is set to T.}
+Entities that have no letters are removed, if remove_nums is set to TRUE.}
 \usage{
-clean_entities(v, remove_nums = T, remove_trailing_s = T, concatenator = "_")
+clean_entities(
+  v,
+  remove_nums = TRUE,
+  remove_trailing_s = TRUE,
+  concatenator = "_"
+)
 }
 \arguments{
 \item{v}{a vector of entity names}
 
-\item{remove_nums}{A boolean. If T, sets entities that contain no letters to an empty string. If F, sets entities that contain no letters or numbers to an empty string.}
+\item{remove_nums}{A boolean. If TRUE, sets entities that contain no letters to an empty string. If FALSE, sets entities that contain no letters or numbers to an empty string.}
 
-\item{remove_trailing_s}{A boolean. If T, removes trailing instances of concatenator+s, or apostrophe+s in each element of v.}
+\item{remove_trailing_s}{A boolean. If TRUE, removes trailing instances of concatenator+s, or apostrophe+s in each element of v.}
 
 \item{concatenator}{Defaults to an underscore. Use regex notation. The concatenator used in elements of v.}
 }
@@ -31,5 +36,5 @@ Any trailing "'s" at the end of the entity name is removed.
 All non-word characters are removed.
 Consecutive underscores are collapsed to a single underscore.
 Leading and trailing underscores are removed.
-Entities that have no letters are removed, if remove_nums is set to T.
+Entities that have no letters are removed, if remove_nums is set to TRUE.
 }
