+++ NAMESPACE
@@ -44,13 +44,13 @@ importFrom(igraph,tail_of)
 importFrom(igraph,vertex_attr)
 importFrom(magrittr,"%>%")
 importFrom(methods,is)
-importFrom(rlang,.data)
 importFrom(network,network)
 importFrom(network,network.density)
 importFrom(network,network.edgecount)
 importFrom(network,network.size)
 importFrom(ohenery,normalize)
 importFrom(pbapply,pblapply)
+importFrom(rlang,.data)
 importFrom(sna,centralization)
 importFrom(sna,connectedness)
 importFrom(sna,gtrans)
