### 2026-04-09 ad088d0
Python helper (#26)
* new find python method

- also updated igraph names to remove deprecated names from old version

* package hygene

- test procedures out of vignette
- load packaged data don't re-parse
- vignette html not in tarball

* remove old data shrink tar

- removed old data used for piloting
- created data-raw/ (buildignored) for helper functions creating the eng_words.rda and verb rda files

* vignette re-org

separate /paper subdirectory
/paper and /vignettes have unique .bib files
old data cleanup

* (auto) Paper PDF Draft

* typo replacements

* minor typo corrections + check 
[…]
