+++ R/textnet_extract.R
@@ -1,16 +1,16 @@
 # Exported function
 # textnet_extract 
 
-#' Takes a parsed spacy document and uses dependency parsing to generate an edgelist, nodelist, verblist, and apposititve list
+#' Takes a parsed spacy document and uses dependency parsing to generate an edgelist, nodelist, verblist, and appositive list
 #'
 #' @param x parsed spacy document
 #' @param concatenator how entity parts are concatenated (defaults to "_")
-#' @param file location where an list object with an edgelist, nodelist, verblist, and appositivelist should be saved as .RDS file
+#' @param file location where a list object with an edgelist, nodelist, verblist, and appositivelist should be saved as .RDS file
 #' @param cl number of cores to crawl sentences in parallel (defaults to 1)
 #' @param keep_entities character vector of spacy entity types to retain, defaults to people (PERSON), organizations (ORG), and geographic entities (GPE)
 #' @param return_to_memory boolean for whether function should return final result as workspace object
-#' @param keep_incomplete_edges Boolean. If T, keeps edges with only a source or target but not both
-#' @param remove_neg Boolean. If T, removes edges whose head token has a negation child
+#' @param keep_incomplete_edges Boolean. If TRUE, keeps edges with only a source or target but not both
+#' @param remove_neg Boolean. If TRUE, removes edges whose head token has a negation child
 #' @return A list with four objects:
 #' \itemize{
 #'    \item nodelist -- data.table of nodes and their attributes
@@ -31,10 +31,10 @@
 #'      \item xcomp_verb -- list of any open causal complements in the verb phrase
 #'      \item xcomp_helper_token -- list of any auxiliary verbs associated with the open causal complements
 #'      \item xcomp_helper_lemma -- list of base forms of any auxiliary verbs associated with the open causal complements
-#'      \item neg -- Boolean: T if there is a negation token present in the verb phrase; F otherwise
-#'      \item edgeiscomplete -- Boolean: T if the edge has both a source and target node; F otherwise
-#'      \item has_hedge -- Boolean: T if the edge contains a verb or auxiliary verb indicating uncertainty, namely "may","might","can","could", "seem","appear","suggest","tend","assume","indicate","doubt", or "believe"; the value is F otherwise
-#'      \item is_future -- Boolean: T if the edge is future tense, as indicated by auxiliary verbs of the form 'is going to' or 'will/shall'; F otherwise 
+#'      \item neg -- Boolean: TRUE if there is a negation token present in the verb phrase; FALSE otherwise
+#'      \item edgeiscomplete -- Boolean: TRUE if the edge has both a source and target node; FALSE otherwise
+#'      \item has_hedge -- Boolean: TRUE if the edge contains a verb or auxiliary verb indicating uncertainty, namely "may","might","can","could", "seem","appear","suggest","tend","assume","indicate","doubt", or "believe"; the value is FALSE otherwise
+#'      \item is_future -- Boolean: TRUE if the edge is future tense, as indicated by auxiliary verbs of the form 'is going to' or 'will/shall'; FALSE otherwise 
 #'      \item doc_sent_verb -- unique ID for the edge indicating the document, sentence, and edge verb token separated by underscores
 #'    }
[…]

[…]
