### 2026-04-10 54f938e
incremented version number; removed directedclustering example
