+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: textNet
 Title: Generate semantic networks from unstructured policy documents
-Version: 1.0.4
+Version: 1.0.5
 Authors@R: 
     c( 
     person(given = "Elise",
@@ -41,7 +41,6 @@ Suggests:
     knitr (>= 1.42),
     ggraph,
     pdftools,
-    DirectedClustering,
     findpython,
     ggplot2,
     gridExtra,
