### 2026-04-13 9c2d501
make script for figure 2 in paper.md
separate figures for paper form figures for vignette
