### 2026-04-13 03a874a
Minor edits for review (#27)
* Update README.md

include more detailed instructions and troubleshooting resources for suggested packages

* removed extra plot

* make script for figure 2 in paper.md

separate figures for paper form figures for vignette

* (auto) Paper PDF Draft

---------

Co-authored-by: E.J. Zufall <90878994+ezufall@users.noreply.github.com>
Co-authored-by: tylerandrewscott <tylerandrewscott@users.noreply.github.com>
