+++ README.md
@@ -21,11 +21,17 @@ respective network characteristics. Its flexibility allows for any
 desired entity categories, such as organizations, geopolitical entities,
 dates, or custom-defined categories, to be preserved.
 
-See vignettes/paper.pdf for an overview of the package functionality and potential use cases.  
+See paper/paper.pdf for an overview of the package functionality and potential use cases.  
 
 To demo the package, see vignettes/textNet_vignette_2025.pdf for a reproducible example that transforms raw text data into event networks.
 
 # Installation
+
+Downloading packages from Github requires creating a Personal Access Token (PAT). Instructions here: https://github.com/orgs/community/discussions/140956
+
+For troubleshooting common problems with PAT, see: https://stackoverflow.com/questions/70908295/failed-to-install-unknown-package-from-github 
+
+You will also need RTools, if not already installed: https://cran.r-project.org/bin/windows/Rtools/
 
 Install the `pak` package if you don't already have it:
 
@@ -46,7 +52,7 @@ pak::local_install()
 ```
 
 # Suggested packages
-The primary function, textnet_extract(), can be used without the use of spaCy, if the user prefers to import compatible data from a separate tool. A wrapper of the spacyr package is included for convenience, to conduct preprocessing in-house. Use of this functionality requires installation of the reticulate and spacyr packages, as seen below. Use of the spacyr wrapper also requires installing spaCy and a language model; the 'en_core_web_lg' model is recommended (see below). For assistance, please see the spacyr documentation page: https://spacyr.quanteda.io/reference/spacy_install.html
+The primary function, textnet_extract(), can be used without the use of spaCy, if the user prefers to import compatible data from a separate tool. A wrapper of the spacyr package is included for convenience, to enable preprocessing in-house. Use of this functionality requires installation of the reticulate and spacyr packages, as seen below. Use of the spacyr wrapper 'parse_text' also requires installing spaCy and a language model; the 'en_core_web_lg' model is recommended (see below). 
 
 ```
 pak::pak(c("reticulate", "spacyr"))
@@ -54,9 +60,19 @@ library(spacyr)
 spacy_install()
 spacy_download_langmodel('en_core_web_lg')
 ```
+Users of spaCy may encounter a variety of unique challenges with this process based on their particular computer setup. For assistance and guidance with common spaCy challenges, please see the spacyr documentation page: https://spacyr.quanteda.io/index.html, the spacy_install documentation page: https://spacyr.quanteda.io/reference/spacy_install.html and the spacyr README: https://cran.r-project.org/web/packages/spacyr/readme/README.html. Windows users experiencing challenges with permissions or access may benefit from this forum post: https://stackoverflow.com/questions/56974927/permission-denied-trying-to-run-python-on-windows-10
 
-# Guidelines for Contributors
-See CONTRIBUTING.md for guidelines for working on this package.
+Use of the 'parse_text' wrapper requires providing a path to an appropriate python binary with required modules "spacy" and the desired language model. External support tools exist that help the user locate the path once the modules have been installed, such as findpython: https://cran.r-project.org/web/packages/findpython/index.html
+
+A wrapper of pdftools is also included for convenience. Use of this wrapper 'pdf_clean' requires installation of pdftools, which in turn requires the poppler library. 
+On Ubuntu/Debian: sudo apt-get install libpoppler-cpp-dev
+On macOS: brew install poppler
+On Windows (in Anaconda Prompt): conda install conda-forge::poppler
+For support and troubleshooting on pdftools and/or poppler, please visit the pdftools documentation page: https://www.rdocumentation.org/packages/pdftools/versions/3.7.0/topics/pdftools
[…]
