### 2026-04-13 51e9bbb
parse_text no longer requires ret_path. now uses spacyr defaults instead
