+++ R/parse_text.R
@@ -5,7 +5,6 @@
 #'
 #' Creates an edgelist and nodelist for each document using spaCy's en_core_web_lg model.
 #'
-#' @param ret_path filepath to use for Sys.setenv reticulate python call. Note: Python and miniconda must already be installed.
 #' @param keep_hyph_together Set to true to replace hyphens within a single word with underscores. Defaults to false.
 #' @param phrases_to_concatenate character vector of phrases, in which each element is a string consisting of tokens separated by spaces. These are replaced with their concatenated version in order, from left to right. It is suggested that the most specific phrases, with the most words, are arranged at the left.
 #' @param concatenator This is a character or string that will be used to replace the spaces in the phrases_to_concatenate.
@@ -25,18 +24,13 @@
 #' @importFrom pbapply pblapply
 #' @export
 
-parse_text <- function(ret_path, keep_hyph_together=FALSE, phrases_to_concatenate=NA,
+parse_text <- function(keep_hyph_together=FALSE, phrases_to_concatenate=NA,
                        concatenator="_", text_list, parsed_filenames,
                        overwrite=TRUE, test=FALSE, custom_entities = NULL, entity_ruler_patterns = NULL,
                        ruler_position = c("after", "before"), overwrite_ents = TRUE){
   if(!requireNamespace("spacyr", quietly = TRUE)){
     stop("Package 'spacyr' must be installed to use this function.",
          call.=FALSE)
-  }
-
-  # Input validation
-  if(!is.character(ret_path) || length(ret_path) != 1) {
-    stop("'ret_path' must be a single character string")
   }
 
   if(!is.logical(keep_hyph_together) || length(keep_hyph_together) != 1) {
@@ -95,14 +89,11 @@ parse_text <- function(ret_path, keep_hyph_together=FALSE, phrases_to_concatenat
   # Fix OpenMP library conflict (common on macOS with conda)
   Sys.setenv(KMP_DUPLICATE_LIB_OK = "TRUE")
 
-  # Set up Python environment
-  Sys.setenv(RETICULATE_PYTHON=ret_path)
   if(!requireNamespace("reticulate", quietly = TRUE)){
     stop("Package 'reticulate' must be installed to use this function.",
          call.=FALSE)
   }
-  reticulate::py_config()
-
+  
   tryCatch({
     spacyr::spacy_initialize(model = "en_core_web_lg")
   }, error = function(e){
