### 2026-04-14 5e4ff9d
removed ret_path from vignette to match new parse_text
