+++ vignettes/textNet_vignette_2025.Rmd
@@ -226,7 +226,7 @@ The *ggraph* package [@pedersen_ggraph_2024] has been used to create the two net
 cat("**Note:** The plots in this vignette require the ggraph, ggplot2, and gridExtra packages. Install them to see the visualizations: `install.packages(c('ggraph', 'ggplot2', 'gridExtra'))`")
 ```
 
-```{r plot, message=F, warning=F, eval=has_viz}
+```{r plot, message=F, warning=F, fig.width=5,fig.height=5, fig.retina=2, eval=has_viz}
 library(ggraph)
 #to plot the networks, the same export_to_network function can be used, 
 #but a weighted version of the graph, such that all edges between 
@@ -516,14 +516,11 @@ old_new_text <- textNet::old_new_text
 
 ### Pre-Processing Step II: Parse Text
 
-This is a wrapper for the pre-trained multipurpose NLP model *spaCy* [@honnibal_spacy_2021], which we access through the R package *spacyr* [@benoit_spacyr_2023]. It produces a table that can be fed into the textnet_extract function in the following step. To initialize the session, the user must define the "RETICULATE_PYTHON" path, abbreviated as "ret_path" in *textNet*, as demonstrated in the example below. The page contents processed in the Step 1 must now be specified in vector form in the "pages" argument. To determine which file each page belongs to, the user must specify the file_ids of each page. We have demonstrated how to do this below. The package by default does not preserve hyphenated terms, but rather treats them as separate tokens. This can be adjusted. 
+This is a wrapper for the pre-trained multipurpose NLP model *spaCy* [@honnibal_spacy_2021], which we access through the R package *spacyr* [@benoit_spacyr_2023]. It produces a table that can be fed into the textnet_extract function in the following step. The page contents processed in the Step 1 must now be specified in vector form in the "pages" argument. To determine which file each page belongs to, the user must specify the file_ids of each page. We have demonstrated how to do this below. The package by default does not preserve hyphenated terms, but rather treats them as separate tokens. This can be adjusted. 
 
 The user may also specify "phrases_to_concatenate", an argument representing a set of phrases for spaCy to keep together during its parsing. The example below demonstrates how to use this feature to supplement the NER capabilities of spaCy with a custom list of entities. This supplementation could be used to ensure that specific known entities are recognized; for instance, spaCy might not detect that a consulting firm such as "Schmidt and Associates" is one entity rather than two. Conversely, this capability could be leveraged to create a new category of entities to detect, that a pretrained model is not designed to specifically recognize. For instance, to create a public health network, one might include a known list of contaminants and diseases and designate custom entity type tags for them, such as "CONTAM" and "DISEASE"). In this example, we investigate the connections between the organizations, people, and geopolitical entities discussed in the plan with the flow of water in the basin. To assist with this, we have input a custom list of known water bodies in the region governed by our test document and have given it the entity designation "WATER". This is carried out by setting the variable "phrases_to_concatenate" to a character vector, including all of the custom entities. Then, the entity type can be set to the desired category. Note that this function is case-sensitive.
 
 ```{r parse, eval = FALSE}
-library(findpython)
-#find the version of python that has the required NLP modules
-ret_path <- find_python_cmd(required_modules = c('spacy', 'en_core_web_lg'))
 
 #example of how to define a custom list for a user-defined entity type
 water_bodies <- c("surface water", "Surface water", "groundwater", "Groundwater",
@@ -544,7 +541,7 @@ if(!requireNamespace("spacyr", quietly = T)){
 #this is the equivalent of the textNet::old_new_parsed sample data
 #that travels with the package
 #it is a thin wrapper for spacyr::spacy_parse
-old_new_parsed <- textNet::parse_text(ret_path,
[…]
