### 2026-04-14 cb02715
Removes outdated reference to arguments in parse_text
