### 2026-04-14 f8ca40c
Parse text mod (#28)
* parse_text no longer requires ret_path. now uses spacyr defaults instead

* removed ret_path from vignette to match new parse_text

* Removes outdated reference to arguments in parse_text

* new parse_text syntax description included in README

---------

Co-authored-by: E.J. Zufall <90878994+ezufall@users.noreply.github.com>
