### 2026-04-14 #7615 @tylerandrewscott
@kbenoit , my apologies, I was working on a new feature to enable using spaCy's en_core_web_trf model on a GPU, and I messed up the testing procedures. We resolved this issue, and also took the opportunity to improve a few more things:
- adjusted vignette build; now only uses required functionality, not optional functionality
- updated test suite, and included instructions in README about additional requirements for testing
- expanded guidance for wrapper support in README
- cleaned up minor typos and updated a few deprecated function names
- streamlined internal data management
- conducted testing of these new protocols on multiple systems: Mac OS, Windows, and Linux
- parse_text uses default spacyr workflow now for accessing spacy/spacy models

Also, for convenience here is your original checklist https://github.com/openjournals/joss-reviews/issues/7615#issuecomment-3484535339
