### 2026-04-28 b737746
add acronym section integration
extract_front_matter_acronyms.R finds acronym sections and grabs the pairs. I can be optionally integrated into find_acronyms.R to combine acronym section and in-text definition
