+++ R/extract_front_matter_acronyms.R
@@ -1,0 +1,175 @@
+#' Extract acronym definitions from a front-matter acronym table.
+#'
+#' Scans a character vector (one element per page or chunk) for a section
+#' heading that signals an explicit acronym/abbreviation table (e.g.,
+#' "Acronyms and Abbreviations", "List of Acronyms"). When found, pastes
+#' that page and up to five following pages, then extracts rows of the form
+#' \code{ACRONYM <separator> Full Name}.
+#'
+#' This is the highest-precision source for acronym definitions because the
+#' document authors have already resolved any ambiguity. It is called
+#' automatically by \code{\link{find_acronyms}} when
+#' \code{use_front_matter = TRUE}.
+#'
+#' @param str A character vector, typically one element per page of the
+#'   document (as returned by \code{pdf_clean}).
+#'
+#' @return A \code{data.table} with columns \code{name} (full expansion,
+#'   spaces and hyphens replaced with underscores to match spaCy entity
+#'   format) and \code{acronym}. Returns an empty \code{data.table} if no
+#'   front-matter acronym table is detected.
+#'
+#' @import data.table
+#' @importFrom stringr str_match_all str_remove str_trim str_replace_all
+#'
+#' @export
+
+# Check that the letter-only portion of `acronym` is a subsequence of the
+# first letters of the words in `expansion`. This filters out rows where the
+# regex matched a stray line of text rather than a real acronym entry.
+#
+# Example: "EPA" vs "Environmental Protection Agency"
+#   acr_letters = "epa"
+#   word initials = e, p, a  ->  "epa"   -> "epa" is a subsequence -> TRUE
+#
+# Example: "DWR" vs "Department of Water Resources"
+#   acr_letters = "dwr"
+#   word initials = d, o, w, r -> "dowr"  -> "dwr" subsequence (skip "o") -> TRUE
+#
+# Non-letter characters in the acronym (digits, &, /) are stripped before
+# matching, so "SB 1383" -> "sb" checked against "Senate Bill ..." -> TRUE.
+.acronym_fits_expansion <- function(acronym, expansion) {
+  acr_letters <- tolower(gsub("[^A-Za-z]", "", acronym))
+  if (nchar(acr_letters) == 0L) return(TRUE)
+
+  words    <- unlist(strsplit(expansion, "\\s+"))
+  initials <- tolower(substr(words, 1L, 1L))
+  initials <- initials[grepl("[a-z]", initials, perl = TRUE)]
+  if (length(initials) == 0L) return(FALSE)
+
+  ini_str   <- paste(initials, collapse = "")
+  acr_chars <- strsplit(acr_letters, "")[[1]]
+
+  # Walk through acr_chars in order; each must be found in ini_str at or
+  # after the current position (subsequence check).
+  pos <- 1L
+  for (ch in acr_chars) {
+    hit <- regexpr(ch, substring(ini_str, pos), fixed = TRUE)
+    if (hit < 0L) return(FALSE)
+    pos <- pos + hit  # advance past the matched character
+  }
+  TRUE
+}
+
+extract_front_matter_acronyms <- function(str) {
+  if (!is.character(str)) {
+    stop("'str' must be a character vector")
+  }
+
+  # Section headings that signal an explicit acronym table.
+  # Anchored to the full line (after optional leading section numbers like
+  # "3.2" or "Appendix A") so mid-paragraph uses of the word "acronyms"
+  # do not trigger a false match.
+  heading_re <- paste0(
+    "(?im)",                          # case-insensitive, multiline
[…]
