### 2026-04-29 e6741cc
Revert "add acronym section integration"
This reverts commit b7377463474dc2202ad991f4c2b0b11ca5bf279e.
