+++ R/find_acronyms.R
@@ -1,47 +1,26 @@
 # Exported function
-# find_acronyms
+# pdf_clean
 
-#' Detect acronym definitions in a character vector.
+#' Take a character vector and detect all parenthetical statements in which an acronym is defined within the character vector. 
 #'
-#' Combines two sources:
-#' \enumerate{
-#'   \item \strong{Inline parenthetical patterns} -- matches like
-#'     \code{"Long Form (ACRONYM)"} or \code{"ACRONYM (Long Form)"}
-#'     anywhere in the text.
-#'   \item \strong{Front-matter acronym table} (when
-#'     \code{use_front_matter = TRUE}) -- an explicit two-column table under
-#'     a heading such as "Acronyms and Abbreviations". These definitions are
-#'     treated as authoritative and win when the two sources conflict.
-#' }
-#'
-#' @param str A character vector, typically one element per page (as returned
-#'   by \code{pdf_clean}).
-#' @param use_front_matter Logical. When \code{TRUE} (default),
-#'   \code{\link{extract_front_matter_acronyms}} is called and its results
-#'   are merged with the inline results, with front-matter entries taking
-#'   precedence on conflict.
-#'
-#' @return A \code{data.table} with columns \code{name} (full expansion,
-#'   spaces and hyphens replaced with underscores) and \code{acronym}.
-#'   Only acronyms with a unique expansion are returned. Rows are sorted
-#'   shortest to longest acronym so that \code{\link{disambiguate}} applies
-#'   shorter substitutions first.
-#'
+#' @param str A character vector
+#' 
 #' @import data.table
 #' @importFrom stringr str_split str_remove_all str_replace_all
 #' @importFrom stringi stri_match_last stri_match_all
-#'
-#' @export
 
-find_acronyms <- function(str, use_front_matter = TRUE){
+#' @return a data table with a "name" column and an "acronym" column representing its associated acronym.
+#' Each row corresponds to a unique match in the document.
+#' 
+#' @export
+#' 
+
+find_acronyms <- function(str){
   # Input validation
   if(!is.character(str)) {
     stop("'str' must be a character vector")
   }
 
-  # ------------------------------------------------------------------
-  # Section 1: Inline parenthetical detection (existing logic)
-  # ------------------------------------------------------------------
   paren_splits <- str_split(str, pattern = "\\)")
   paren_splits2 <- lapply(paren_splits, function (k) k[nchar(k)>0])
   paren_splits3 <- lapply(paren_splits2, function(m) stringr::str_split(m, pattern = "\\("))
@@ -51,67 +30,46 @@ find_acronyms <- function(str, use_front_matter = TRUE){
   paren_splits$acr1 <- stri_match_last(str = paren_splits5[,1], regex ="\\b[A-Z]+\\b")
   paren_splits$acr2 <- stri_match_all(str = paren_splits5[,2], regex ="\\b[A-Z]+\\b")
   paren_splits$abb1 <- str_remove_all(paren_splits5[,1],"[^A-Z]")
-  paren_splits$abb1 <- sapply(1:length(paren_splits$acr2),
+  paren_splits$abb1 <- sapply(1:length(paren_splits$acr2), 
                                   function(s) sapply(1:length(paren_splits$acr2[[s]]), function (m){
-                                    stri_match_last(str=paren_splits$abb1[s],
+                                    stri_match_last(str=paren_splits$abb1[s], 
                                         regex = paren_splits$acr2[[s]][m])
                                   }))
   sp_lower <- "[\\s|a-z]+"
-  paren_splits$name1 <- sapply(1:length(paren_splits$acr2),
+  paren_splits$name1 <- sapply(1:length(paren_splits$acr2), 
[…]
