### 2026-05-14 #7615 @kbenoit
I was unclear on that too. I seem to get a lot of mail from threads on JOSS papers for which I was once mentioned but am not an assignee. Was not sure what action is needed, if any, from me here. While using GitHub issues is a kinda cool way to manage a peer review process, it has some drawbacks when it comes to clarity of who is supposed to do what. @crvernon please advise. @ezufall Apologies if my misderstanding of the process has cause delays, as I know this review has taken forever.
