### 2026-05-29 ec7c63f
Rename find_acronyms() to find_intext_acronyms() with deprecation alias
find_acronyms() is preserved as a deprecated alias that emits
.Deprecated() and forwards to find_intext_acronyms(). The rename
clarifies the function's scope (parenthetical acronyms found
inline in the text) in anticipation of a complementary
extract_front_matter_acronyms() that operates on front-matter
acronym tables.
