+++ R/find_intext_acronyms.R
@@ -1,21 +1,45 @@
 # Exported function
-# find_acronyms
+# find_intext_acronyms
 
-#' Take a character vector and detect all parenthetical statements in which an acronym is defined within the character vector. 
+#' Detect parenthetically-defined acronyms in a character vector.
 #'
-#' @param str A character vector
-#' 
+#' Scans a character vector for inline parenthetical patterns of the form
+#' \code{"Long Form (ACRONYM)"} or \code{"ACRONYM (Long Form)"} and returns
+#' the matches as a two-column \code{data.table}. Pair the result with
+#' \code{\link{disambiguate}} to rename acronym mentions in a textnet
+#' extract to their full forms.
+#'
+#' For an alternative source of acronym definitions (the front-matter
+#' acronym table that many policy documents include), see
+#' \code{\link{extract_front_matter_acronyms}}. The two functions return
+#' compatible columns and can be combined with \code{rbind()}.
+#'
+#' @param str A character vector.
+#'
+#' @return A \code{data.table} with columns \code{name} (full expansion,
+#'   spaces and hyphens replaced with underscores so the name matches
+#'   spaCy's concatenated entity tokens) and \code{acronym}. Each row is a
+#'   unique acronym; rows where the same acronym appears with multiple
+#'   inline expansions are dropped to avoid ambiguous downstream
+#'   replacements. Rows are sorted shortest to longest acronym so
+#'   replacements in \code{\link{disambiguate}} apply in the right order.
+#'
 #' @import data.table
 #' @importFrom stringr str_split str_remove_all str_replace_all
 #' @importFrom stringi stri_match_last stri_match_all
-
-#' @return a data table with a "name" column and an "acronym" column representing its associated acronym.
-#' Each row corresponds to a unique match in the document.
-#' 
+#'
+#' @examples
+#' # Inline parenthetical patterns are detected anywhere in the text.
+#' text <- c(
+#'   "The Department of Water Resources (DWR) coordinates with the State",
+#'   "Water Resources Control Board (SWRCB) on groundwater policy.",
+#'   "Groundwater Sustainability Plans (GSPs) are submitted annually."
+#' )
+#' find_intext_acronyms(text)
+#'
 #' @export
-#' 
 
-find_acronyms <- function(str){
+find_intext_acronyms <- function(str){
   # Input validation
   if(!is.character(str)) {
     stop("'str' must be a character vector")
@@ -30,38 +54,38 @@ find_acronyms <- function(str){
   paren_splits$acr1 <- stri_match_last(str = paren_splits5[,1], regex ="\\b[A-Z]+\\b")
   paren_splits$acr2 <- stri_match_all(str = paren_splits5[,2], regex ="\\b[A-Z]+\\b")
   paren_splits$abb1 <- str_remove_all(paren_splits5[,1],"[^A-Z]")
-  paren_splits$abb1 <- sapply(1:length(paren_splits$acr2), 
+  paren_splits$abb1 <- sapply(1:length(paren_splits$acr2),
                                   function(s) sapply(1:length(paren_splits$acr2[[s]]), function (m){
-                                    stri_match_last(str=paren_splits$abb1[s], 
+                                    stri_match_last(str=paren_splits$abb1[s],
                                         regex = paren_splits$acr2[[s]][m])
                                   }))
   sp_lower <- "[\\s|a-z]+"
-  paren_splits$name1 <- sapply(1:length(paren_splits$acr2), 
+  paren_splits$name1 <- sapply(1:length(paren_splits$acr2),
[…]
