+++ man/find_intext_acronyms.Rd
@@ -1,0 +1,50 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/find_intext_acronyms.R
+\name{find_intext_acronyms}
+\alias{find_intext_acronyms}
+\alias{find_acronyms}
+\title{Detect parenthetically-defined acronyms in a character vector.}
+\usage{
+find_intext_acronyms(str)
+
+find_acronyms(str)
+}
+\arguments{
+\item{str}{A character vector.}
+}
+\value{
+A \code{data.table} with columns \code{name} (full expansion,
+  spaces and hyphens replaced with underscores so the name matches
+  spaCy's concatenated entity tokens) and \code{acronym}. Each row is a
+  unique acronym; rows where the same acronym appears with multiple
+  inline expansions are dropped to avoid ambiguous downstream
+  replacements. Rows are sorted shortest to longest acronym so
+  replacements in \code{\link{disambiguate}} apply in the right order.
+}
+\description{
+Scans a character vector for inline parenthetical patterns of the form
+\code{"Long Form (ACRONYM)"} or \code{"ACRONYM (Long Form)"} and returns
+the matches as a two-column \code{data.table}. Pair the result with
+\code{\link{disambiguate}} to rename acronym mentions in a textnet
+extract to their full forms.
+}
+\details{
+For an alternative source of acronym definitions (the front-matter
+acronym table that many policy documents include), see
+\code{\link{extract_front_matter_acronyms}}. The two functions return
+compatible columns and can be combined with \code{rbind()}.
+
+\code{find_acronyms} is a deprecated alias for \code{find_intext_acronyms}
+kept for backward compatibility with code written against textNet
+\eqn{\le}{<=} 1.0.1. It emits a deprecation warning and forwards to the
+new name.
+}
+\examples{
+# Inline parenthetical patterns are detected anywhere in the text.
+text <- c(
+  "The Department of Water Resources (DWR) coordinates with the State",
+  "Water Resources Control Board (SWRCB) on groundwater policy.",
+  "Groundwater Sustainability Plans (GSPs) are submitted annually."
+)
+find_intext_acronyms(text)
+}
