+++ NAMESPACE
@@ -8,6 +8,7 @@ export(entity_specify)
 export(export_to_network)
 export(filter_sentences)
 export(find_acronyms)
+export(find_intext_acronyms)
 export(parse_text)
 export(parse_text_trf)
 export(pdf_clean)
