### 2026-05-29 f5ea4be
Add extract_front_matter_acronyms()
Extracts acronym definitions from explicit 'Acronyms and
Abbreviations' / 'List of Acronyms' tables at the front of
a document. Returns the same (name, acronym) data.table shape
as find_intext_acronyms() so the two can be rbind()'d and the
combined table fed into disambiguate().

Handles Title Case nicknames (District, Authority, Mather AFB),
mixed-case acronyms (InSAR, LiDAR, C2VSimCG), and both
dot-leader and wide-whitespace separator styles. All heading
occurrences are processed, so per-chapter acronym tables are
fully harvested. A post-match blacklist fi
[…]
