+++ man/extract_front_matter_acronyms.Rd
@@ -1,0 +1,73 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extract_front_matter_acronyms.R
+\name{extract_front_matter_acronyms}
+\alias{extract_front_matter_acronyms}
+\title{Extract acronym definitions from a front-matter acronym table.}
+\usage{
+extract_front_matter_acronyms(str, max_window_pages = 10L)
+}
+\arguments{
+\item{str}{A character vector, typically one element per page of the
+document (as returned by \code{\link{pdf_clean}}).}
+
+\item{max_window_pages}{Integer. Upper bound on how many pages after
+each heading the function will read (defaults to 10). The actual window
+stops earlier if two consecutive pages contain no table-shaped rows or
+if the next heading page is reached.}
+}
+\value{
+A \code{data.table} with columns \code{name} (full expansion,
+  spaces and hyphens replaced with underscores to match spaCy's
+  concatenated entity tokens) and \code{acronym}. Returns an empty
+  \code{data.table} (zero rows, both columns) when no front-matter
+  acronym table is detected.
+}
+\description{
+Scans a character vector (typically one element per page, as returned by
+\code{\link{pdf_clean}}) for one or more section headings that signal an
+explicit acronym / abbreviation table (e.g., "Acronyms and Abbreviations",
+"List of Acronyms", or just the bare word "Abbreviations"). For each
+heading found, the function walks the heading page and subsequent pages
+while they continue to contain table-shaped rows, and extracts rows of
+the form \code{ACRONYM <separator> Full Name}.
+}
+\details{
+All heading occurrences in the document are processed, so a document
+with separate "Abbreviations" and "Acronyms" sections (or per-chapter
+acronym tables) is fully harvested. Conflicting acronym definitions are
+resolved by keeping the first occurrence.
+
+The output has the same column shape as \code{\link{find_intext_acronyms}},
+so the two can be \code{rbind()}'d (front-matter rows first, so
+front-matter wins on conflict after \code{unique(by = "acronym")}) and the
+combined table fed into \code{\link{disambiguate}}.
+
+By default only capital-letter-leading acronyms are captured. Lowercase
+technical units (e.g., \code{bgs}, \code{cfs}, \code{mg/L}) are skipped
+because they rarely correspond to entity nodes and admitting them
+substantially increases the false-positive surface in justified-column
+PDFs. Title Case "nickname" acronyms (\code{District}, \code{Authority},
+\code{Mather AFB}) \emph{are} captured.
+}
+\examples{
+# Input is a character vector with one element per page, as returned by
+# pdf_clean(). Here we hand-build a minimal two-page example.
+pages <- c(
+  paste(
+    "Acronyms and Abbreviations",
+    "",
+    "EPA          Environmental Protection Agency",
+    "DWR          Department of Water Resources",
+    "GSA          Groundwater Sustainability Agency",
+    "GSP          Groundwater Sustainability Plan",
+    sep = "\n"
+  ),
+  "(next page of the document; no acronyms here)"
+)
+extract_front_matter_acronyms(pages)
+
+# Combine with find_intext_acronyms() so front-matter rows win on conflict.
+fm <- extract_front_matter_acronyms(pages)
+inline <- find_intext_acronyms(pages)
+all_acronyms <- unique(rbind(fm, inline), by = "acronym")
+}
