+++ NAMESPACE
@@ -6,6 +6,7 @@ export(combine_networks)
 export(disambiguate)
 export(entity_specify)
 export(export_to_network)
+export(extract_front_matter_acronyms)
 export(filter_sentences)
 export(find_acronyms)
 export(find_intext_acronyms)
@@ -67,10 +68,12 @@ importFrom(stringi,stri_match_last)
 importFrom(stringi,stri_read_lines)
 importFrom(stringi,stri_replace_all_regex)
 importFrom(stringr,str_detect)
+importFrom(stringr,str_match_all)
 importFrom(stringr,str_remove)
 importFrom(stringr,str_remove_all)
 importFrom(stringr,str_replace_all)
 importFrom(stringr,str_split)
+importFrom(stringr,str_trim)
 importFrom(tidyr,expand)
 importFrom(tidyr,tibble)
 importFrom(tools,file_ext)
