### 2026-05-29 295dd8f
Add tests + fixtures for extract_front_matter_acronyms; NEWS entry
Adds five regression fixtures under tests/testthat/testdata/
acronym_pages/ (form-feed-separated page text from real GSP
PDFs) with hand-validated ground-truth CSVs totaling 438
expected (name, acronym) rows. Tests pin function output
against the truth CSVs, plus a negative case (narrative text
returns empty data.table) and input-validation checks.

NEWS.md seeded with a 'development version' entry. The
version bump on DESCRIPTION is deferred until merge time so
this branch does not preempt the version assignment for the
make_t
[…]
