### 2026-05-31 0ce7154
build for running with dual acronym detectors
