+++ DESCRIPTION
@@ -19,7 +19,6 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.3
 Imports:
     magrittr,
     dplyr,
@@ -54,3 +53,4 @@ Suggests:
 VignetteBuilder: knitr
 Depends: 
     R (>= 3.5.0)
+Config/roxygen2/version: 8.0.0
