### 2026-07-06 2ca4acd
claude pass at modernizing descriptions
