### 2026-07-06 f586045
split pdf/text/parse example data into v1 / v2 with more descriptive names
