+++ tests/testthat/test-textnet_extract.R
@@ -5,11 +5,15 @@ library(stringr)
 
 # Use the bundled parsed data (equivalent to parse_text() output on the
 # extdata PDFs; avoids spaCy dependency in tests)
-old_new_parsed <- textNet::old_new_parsed
 
 # water_bodies and ent_types come from helper-fixtures.R
 
-extracts <- lapply(old_new_parsed, function(parsed) {
+parsed_versions <- list(
+  v1 = textNet::example_plan_v1_parsed,
+  v2 = textNet::example_plan_v2_parsed
+)
+
+extracts <- lapply(parsed_versions, function(parsed) {
   textnet_extract(parsed, cl = 2,
                   keep_entities = ent_types,
                   keep_incomplete_edges = TRUE)
@@ -18,9 +22,9 @@ extracts <- lapply(old_new_parsed, function(parsed) {
 # For each document, all node entities should be recoverable from the
 # raw parsed token stream. (Appositives may cause a node entity to be a
 # substring of the original, so we use regex substring matching.)
-for (m in seq_along(old_new_parsed)) {
+for (doc_label in names(parsed_versions)) {
 
-  onp <- old_new_parsed[[m]] |>
+  onp <- parsed_versions[[doc_label]] |>
     dplyr::mutate(entitynum = cumsum(str_detect(entity, "_B")))
   onp$entitynum <- ifelse(onp$entity == "", NA, onp$entitynum)
   onp <- onp |>
@@ -39,11 +43,8 @@ for (m in seq_along(old_new_parsed)) {
     remove_nums
   )))
 
-  nodentities <- unique(sort(extracts[[m]]$nodelist$entity_name)) |>
+  nodentities <- unique(sort(extracts[[doc_label]]$nodelist$entity_name)) |>
     str_replace_all("_", "_.*_*")
-
-  doc_label <- names(old_new_parsed)[m]
-  if (is.null(doc_label)) doc_label <- paste("document", m)
 
   test_that(paste("all node entities found in parsed token stream:", doc_label), {
     expect_true(
