### 2026-07-06 1fea4de
consolidate some duplicative code into helper functions
