### 2026-07-06 0e7b2c4
label/name check fixes
