### 2026-07-06 914203c
fix disambiguate() documentation
