+++ R/disambiguate.R
@@ -1,45 +1,5 @@
-# Exported functions 
-# disambiguate 
-
-#' @title Rename entities in a textnet_extract object
-#' @description
-#' Replaces entity names supplied in \code{from} with the corresponding names
-#' in \code{to} throughout a [textnet_extract()] output. Supports partial
-#' matching, prefix removal, and recursive chaining of replacements.
-#'
-#' @param from A list of character vectors representing terms to look for, the same length as 'to'. 
-#' If a term in this list is found, this function replaces it with its corresponding match in 'to'. 
-#' The "acronyms" column from the find_acronyms output could be used here. 
-#' @param to A list of character vectors representing replacement terms, the same length as 'from'. 
-#' If a term in the 'from' list is found, it is replaced with the corresponding term in this list. 
-#' The "names" column in the find_acronyms output could be used here. If an element consists of multiple entities in a vector, 
-#' the relevant edges are duplicated in the edgelist, with one edge for each entity in the vector. 
-#' These vector entities are disambiguated first, followed by the rest of the strings in order of appearance in 'from' and 'to'.
-#rows such that there is an edge for each of the entities in the "to" cell. 
-#' @param match_partial_entity A logical vector of the same length as 'from'. 
-#' If match_partial_entity is TRUE for an element, it can match on the 'from' term separated by concatenator. 
-#' Otherwise, the 'from' term must match the whole entity name to be accepted. Defaults to "FALSE" for all elements.
-#' @param textnet_extract An output of the function textnet_extract
-#' @param try_drop A regex expression representing one or more terms to try dropping. 
-#' The usual case for this is country or state names, such as "^US_". 
-#' If try_drop is supplied, the function first tries to determine whether each element in the textnet_extract matches a term in 'from.'
-#' If not, it determines whether removing try_drop from the remaining elements enables them to match a term in 'from'. 
-#' If so, the matching textnet_extract elements are replaced with the corresponding element in 'to.' 
-#' The non-matches remains unchanged.
-#' @param recursive A logical value, defaulting to TRUE. If recursive is TRUE, the function runs multiple times. 
-#' The number of times the function is run is determined by the longest "chain" in which a value in 'to' is found in 'from', 
-#' which may in turn correspond to a 'to' value that is found in 'from', and so on. 
-#' @param concatenator The word boundary to look for when match_partial_entity is true. Defaults to "_". 
-#' @return a cleaned textNet extract. See textnet_extract help file for structure.
-#' 
-#' @import data.table
-#' @importFrom stringr str_detect str_remove_all str_replace_all
-#' @importFrom dplyr arrange desc filter
-#' @importFrom rlang .data
-#' @importFrom magrittr %>%
-#' @importFrom methods is
-#' @export
-#'
+# Exported functions
+# disambiguate
 
 # Helper: compute how many recursive passes are needed and whether a chain is infinite
 #' @keywords internal
@@ -62,6 +22,44 @@ count_chain_length <- function(to, from) {
   list(times_to_repeat = times_to_repeat, is_inf_loop = is_inf_loop, carryovers = carryovers)
 }
 
+#' @title Rename entities in a textnet_extract object
+#' @description
+#' Replaces entity names supplied in \code{from} with the corresponding names
+#' in \code{to} throughout a [textnet_extract()] output. Supports partial
+#' matching, prefix removal, and recursive chaining of replacements.
+#'
+#' @param from A list of character vectors representing terms to look for, the same length as 'to'.
+#' If a term in this list is found, this function replaces it with its corresponding match in 'to'.
+#' The "acronyms" column from the find_acronyms output could be used here.
+#' @param to A list of character vectors representing replacement terms, the same length as 'from'.
[…]
