+++ man/clean_entities.Rd
@@ -24,7 +24,7 @@ clean_entities(
 a cleaned vector of entity names
 }
 \description{
-Normalises a character vector of entity names by converting mathematical
+Normalizes a character vector of entity names by converting mathematical
 Unicode characters to ASCII letters, removing non-word characters,
 collapsing consecutive concatenators, and stripping leading/trailing
 concatenators.
