### 2026-07-06 fe2e546
fix a couple more long titles
