+++ R/entity_consolidate_replicate.R
@@ -1,10 +1,17 @@
 # Exported functions 
 # entity_consolidate_replicate 
 
-#' Modified version of spacyr::entity_consolidate.
-#' spacyr::entity_consolidate returns a data.frame that no longer has head_token_id and dep_rel data.
-#' Instead, this function basically breaks off a feature of the original function and adds a new column to the original spacyr data.frame that is the concatenated entity.
-#' This feature is inefficient in that the concatenated entity is then replicated multiple times, but this does seem to be the easiest way to preserve the other data.
+#' @title Consolidate named entities while preserving dependency data
+#' @description
+#' Adds a concatenated-entity column to a parsed spacy data frame, keeping the
+#' token-level dependency information that [spacyr::entity_consolidate()] drops.
+#'
+#' @details
+#' Unlike [spacyr::entity_consolidate()], which returns a data frame that no
+#' longer retains `head_token_id` and `dep_rel`, this function adds a new column
+#' holding the concatenated entity while leaving the original annotations intact.
+#' The concatenated entity is replicated across every token of the entity, which
+#' is inefficient but preserves the remaining columns.
 #'
 #' @param x parsed spacy document in data.frame format
 #' @param concatenator A character that separates string segments when they are collapsed into a single entity. Defaults to "_"
