### 2026-07-06 c11fd69
put all data in one spot
