+++ R/textNet-data.R
@@ -1,0 +1,216 @@
+## Package datasets
+## Documentation for all data objects shipped with textNet.
+
+# ---- Reference data --------------------------------------------------------
+
+#' @title English word list from SCOWL
+#' @description
+#' A character vector of English words drawn from the SCOWL 2020.12.07
+#' word list compiled by Kevin Atkinson.
+#'
+#' @details
+#' Includes words from files with extension \eqn{\leq} 60 whose filenames
+#' contain \code{"variant"}, \code{"american"}, \code{"british"},
+#' \code{"canadian"}, or \code{"australian"}. Used by [filter_sentences()] to
+#' remove non-English sentences from parsed text.
+#'
+#' @format A character vector of English words.
+#'
+#' @references
+#' Kevin Atkinson (2020). SCOWL (Spell Checker Oriented Word Lists).
+#'
+#' @source \href{https://sourceforge.net/projects/wordlist/files/SCOWL/2020.12.07/scowl-2020.12.07.zip/download}{SCOWL 2020.12.07}
+#'
+#' @examples
+#' # number of words in the list
+#' length(eng_words)
+#'
+#' # check whether a word is present
+#' "governance" %in% eng_words
+"eng_words"
+
+#' @title VerbNet 3.3 verb classifications
+#' @description
+#' Verb class and type assignments from VerbNet 3.3, used to annotate the
+#' verblist returned by [textnet_extract()].
+#'
+#' @details
+#' Each row corresponds to a verb lemma and its associated VerbNet class and
+#' type. These attributes are joined to the verblist by \code{head_verb_lemma}.
+#' The data were processed from the VerbNet 3.3 source files using
+#' \code{data-raw/verbnet_port.R}.
+#'
+#' @format A \code{data.table} with columns:
+#' \describe{
+#'   \item{head_verb_lemma}{Base form of the verb.}
+#'   \item{classes}{VerbNet 3.3 class membership string.}
+#'   \item{type_name}{VerbNet 3.3 type label.}
+#'   \item{type_id}{Integer identifier for \code{type_name}.}
+#' }
+#'
+#' @references
+#' Schuler, Karin Kipper (2005). \emph{VerbNet: A Broad-Coverage, Comprehensive
+#' Verb Lexicon}. University of Pennsylvania.
+#'
+#' @source \href{https://verbs.colorado.edu/verb-index/vn3.3/}{VerbNet 3.3}
+#'
+#' @examples
+#' # number of verb entries
+#' nrow(verb_classifications)
+#'
+#' # VerbNet types present in the data
+#' unique(verb_classifications$type_name)
+"verb_classifications"
+
+# ---- Example corpus --------------------------------------------------------
+
+#' @title Parsed sample from a groundwater plan corpus
+#' @description
+#' Token-level linguistic annotations for 20 pages drawn from a collection
+#' of publicly available groundwater sustainability plans.
+#'
+#' @details
+#' Parsed with spaCy's \code{en_core_web_lg} model. The data format matches
+#' the output of [parse_text()].
+#'
+#' @format An object of class \code{"spacyr_parsed"}: a data frame containing
+#'   token-level annotations, including document and sentence identifiers,
+#'   tokens, lemmas, part-of-speech tags, dependency relations, and named
+#'   entity labels.
+#'
+#' @examples
+#' # number of tokens
+#' nrow(sample_20p)
+#'
+#' # inspect token, lemma, and part-of-speech annotations
+#' head(sample_20p[, c("token", "lemma", "pos")])
+#'
+#' # most frequent named entity types
+#' head(sort(table(sample_20p$entity_type), decreasing = TRUE), 5)
+"sample_20p"
+
+#' @title Raw text from version 1 of a groundwater plan
+#' @description
[…]
