### 2026-07-17 33ee0e5
restored two versions of while loop to account for partial matches
