### 2026-07-17 2323c50
example improvements
