### 2026-07-17 69a914a
description & example improvements
