+++ DESCRIPTION
@@ -49,3 +49,4 @@ Depends:
     R (>= 3.5.0)
 Config/testthat/edition: 3
 Config/roxygen2/version: 8.0.0
+RoxygenNote: 7.3.3
