### 2026-07-17 45e583b
added WATER vignette reference
