### 2026-08-15 #7615 @ezufall
Hi @crvernon, we have completed the requested changes. We have also confirmed the formatting of the title of the paper in paper.md, conducted a GitHub release of the current reviewed version (tag v1.1.0), and archived this in Zenodo. The DOI from Zenodo for textNet 1.1.0 is 10.5281/zenodo.21944604 (https://doi.org/10.5281/zenodo.21944604)
