+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @software{hollway_manynet_2024,
-	title = {manynet: Many Ways to  Make, Modify, Map, Mark, and Measure Myriad Networks},
+	title = {{m}anynet: Many Ways to  Make, Modify, Map, Mark, and Measure Myriad Networks},
 	url = {https://CRAN.R-project.org/package=manynet},
 	version = {1.2.6},
 	author = {Hollway, James},
@@ -38,8 +38,8 @@
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/psj.12556},
 	doi = {10.1111/psj.12556},
 	abstract = {This paper demonstrates an automated workflow for extracting network data from policy documents. We use natural language processing tools, part-of-speech tagging, and syntactic dependency parsing, to represent relationships between real-world entities based on how they are described in text. Using a corpus of regional groundwater management plans, we demonstrate unique graph motifs created through parsing syntactic relationships and how document-level syntax can be aggregated to develop large-scale graphs. This approach complements and extends existing methods in public management and governance research by (1) expanding the feasible geographic and temporal scope of data collection and (2) allowing for customized representations of governance systems to fit different research applications, particularly by creating graphs with many different node and edge types. We conclude by reflecting on the challenges, limitations, and future directions of automated, text-based methods for governance research.},
-	issue = {941--954},
 	number = {4},
+	pages = {941--954},
 	journaltitle = {Policy Studies Journal},
 	author = {Zufall, Elise and Scott, Tyler A.},
 	urldate = {2024-11-18},
@@ -58,7 +58,7 @@
 }
 
 @software{honnibal_spacy_2021,
-	title = {{spaCy}: Industrial-strength Natural Language Processing in Python},
+	title = {{spaCy}: Industrial-strength Natural Language Processing in {Python}},
 	url = {https://github.com/explosion/spaCy/tree/master},
 	version = {3.1.3},
 	author = {Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane},
@@ -77,33 +77,32 @@
 	date = {2022-10-11},
 }
 
-@software{csardi_pak_2024,
-	title = {pak: Another Approach to Package Installation},
-	rights = {{GPL}-3},
-	url = {https://cran.r-project.org/web/packages/pak/index.html},
+@software{csardi_pak_2026,
+	title = {{pak}: {Another} Approach to Package Installation},
+	copyright = {GPL-3},
 	shorttitle = {pak},
-	abstract = {The goal of pak is to make package installation faster and more reliable. In particular, it performs all HTTP operations in parallel, so metadata resolution and package downloads are fast. Metadata and package files are cached on the local disk as well.},
-	version = {0.8.0},
-	author = {Csárdi, Gábor and Hester, Jim and Software, Posit},
-	urldate = {2025-01-29},
-	date = {2024-09-20},
+	url = {https://cran.r-project.org/web/packages/pak/index.html},
[…]
