### 2026-08-18 #7615 @crvernon
@ezufall note that the title of your archive, as well as the license specified, does not match your paper and the license in your repository.  Please fix these by editing the metadata in your Zenodo archive - no new release required.

Also your references shouldn't be hyperlinks.  They should use the correct bibtex notation that notes the bibtex tag for the appropriate reference.

Let me know when these have been fixed.  Thank you.
