### 2026-08-25 b4fa85d
version 1.1.1 with MIT license
