+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: textNet
 Title: Generate semantic networks from unstructured policy documents
-Version: 1.1.0
+Version: 1.1.1
 Authors@R: 
     c( 
     person(given = "Elise",
