+++ DESCRIPTION
@@ -15,7 +15,6 @@ Authors@R:
 	   )
 Description: 
     textNet is an R package that provides an interface for using spacyr to extract relational data from large policy documents (e.g., plans, rule proposals, impact assessments) and represent these data as a semantic network.
-License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
@@ -49,3 +48,4 @@ Depends:
     R (>= 3.5.0)
 Config/testthat/edition: 3
 Config/roxygen2/version: 8.0.0
+License: MIT + file LICENSE
