### 2026-08-25 #7615 @ezufall
Hello @crvernon, thank you for your patience and for your edits. In the process of implementing the requests we saw that the license needed to be updated. This is now resolved with a new release (tag v1.1.1), and a new version number. Here is the updated Zenodo repository link: https://doi.org/10.5281/zenodo.22100705. The DOI from Zenodo for textNet 1.1.1 is 10.5281/zenodo.22100705. Please let us know if there is anything else we need to resolve with the release metadata.

However, we do not see the hyperlinks you are talking about. Do you mean in the paper proof? The materials for generating our paper proof are in the "paper" folder of our GitHub, including the paper.md file containing the paper text with BibTeX tags for the references, along with a BibTeX file containing the appropriate references. The references show up for me at the end of the paper with correct formatting, not simply hyperlinks. The in-text citations in our paper use author and date, which are automatically generated from the BibTeX notation.
