### 2026-08-25 #7615 @crvernon
🥳 Congratulations on your new publication @ezufall ! Many thanks to @amatsuo, @stuartyeates, @kbenoit, and @ImBIOS for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
